move proper tests to runtime
