+++ projects/orquestra-sdk/tests/runtime/ray/conftest.py
@@ -5,7 +5,7 @@ import os
 
 import pytest
 
-from orquestra.sdk._client._base._testing import _connections
+from orquestra.workflow_runtime._testing import _connections
 
 
 @pytest.fixture(scope="module")
