+++ projects/orquestra-sdk/tests/runtime/ray/test_integration.py
@@ -8,33 +8,28 @@ RuntimeInterface mocks instead of extending this file.
 """
 import json
 import os
-import re
+
 import sys
 import time
 import typing as t
-from datetime import datetime, timezone
+
 from pathlib import Path
 from unittest.mock import ANY, Mock, call, create_autospec
 
 import pytest
-from freezegun import freeze_time
 from orquestra.workflow_runtime._ray import _build_workflow, _client, _dag, _ray_logs
 from orquestra.workflow_runtime._ray._env import (
     RAY_DOWNLOAD_GIT_IMPORTS_ENV,
     RAY_TEMP_PATH_ENV,
 )
 from orquestra.workflow_shared import exceptions
-from orquestra.workflow_shared._graphs import iter_invocations_topologically
 from orquestra.workflow_shared.abc import RuntimeInterface
-from orquestra.workflow_shared.exceptions import OrquestraSDKVersionMismatchWarning
 from orquestra.workflow_shared.schema import ir
 from orquestra.workflow_shared.schema.responses import JSONResult
 from orquestra.workflow_shared.schema.workflow_run import (
     State,
-    WorkflowRun,
     WorkflowRunId,
 )
-from orquestra.workflow_shared.serde import deserialize
 
 from orquestra import sdk
 from orquestra.sdk._client._base._config import LOCAL_RUNTIME_CONFIGURATION
@@ -91,666 +86,6 @@ def _wait_to_finish_wf(run_id: str, runtime: RuntimeInterface, timeout=20.0):
         ]
 
     _poll_loop(_continue_condition, interval=0.1, timeout=timeout)
-
-
-def _count_task_runs(wf_run, state: State) -> int:
-    return len([run for run in wf_run.task_runs if run.status.state == state])
-
-
-@pytest.mark.slow
-class TestRayRuntimeMethods:
-    """
-    Tests that call RayRuntime methods with a real Ray connection and
-    validate that our glue code over Ray works correctly.
-    """
-
-    class TestCreateWorkflowRun:
-        """
-        Tests that validate .create_workflow_run().
-        """
-
-        def test_running_same_workflow_def_twice(self, runtime: _dag.RayRuntime):
-            wf_def = _example_wfs.multioutput_wf.model
-            run_id1 = runtime.create_workflow_run(wf_def, None, False)
-            run_id2 = runtime.create_workflow_run(wf_def, None, False)
-
-            assert run_id1 != run_id2
-
-            _wait_to_finish_wf(run_id1, runtime)
-            _wait_to_finish_wf(run_id2, runtime)
-            output1 = runtime.get_workflow_run_outputs_non_blocking(run_id1)
-            output2 = runtime.get_workflow_run_outputs_non_blocking(run_id2)
-
-            assert output1 == output2
-
-        def test_sets_context(self, runtime: _dag.RayRuntime):
-            wf_def = _example_wfs.wf_with_exec_ctx().model
-
-            # when
-            run_id = runtime.create_workflow_run(wf_def, None, False)
-
-            # then
-            _wait_to_finish_wf(run_id, runtime)
-            outputs = runtime.get_workflow_run_outputs_non_blocking(run_id)
-            assert outputs == (JSONResult(value='"RAY"'),)
-
-        def test_sets_run_id_from_env(
-            self, monkeypatch: pytest.MonkeyPatch, runtime: _dag.RayRuntime
-        ):
-            wf_def = _example_wfs.multioutput_wf.model
-            # Appending the normal run ID to prevent collisions
-            global_wf_run_id = "run_id_from_env" + _dag._generate_wf_run_id(wf_def)
-            monkeypatch.setenv("GLOBAL_WF_RUN_ID", global_wf_run_id)
-
-            run_id = runtime.create_workflow_run(wf_def, None, False)
-
-            runtime.stop_workflow_run(run_id)
-            assert run_id == global_wf_run_id
-
-        def test_simple_workflow_dry_run(
-            self, monkeypatch: pytest.MonkeyPatch, runtime: _dag.RayRuntime
-        ):
-            wf_def = _example_wfs.exception_wf_with_multiple_values.model
-            run_id = runtime.create_workflow_run(wf_def, project=None, dry_run=True)
-            _wait_to_finish_wf(run_id, runtime)
-
-            # normally this WF would fail, but as a dry-run, no task code is executed
-            outputs = runtime.get_workflow_run_outputs_non_blocking(run_id)
-
-            assert outputs == (JSONResult(value='"dry_run task output"'),)
-
-        def test_unpacking_workflow_dry_run(
-            self, monkeypatch: pytest.MonkeyPatch, runtime: _dag.RayRuntime
-        ):
-            wf_def = _example_wfs.multioutput_task_failed_wf.model
-            run_id = runtime.create_workflow_run(wf_def, project=None, dry_run=True)
-            _wait_to_finish_wf(run_id, runtime)
-
-            outputs = runtime.get_workflow_run_outputs_non_blocking(run_id)
-
-            res = JSONResult(value='"dry_run task output"')
-            assert outputs == (
-                res,
-                res,
-                res,
-                res,
-                JSONResult(
-                    value='{"__tuple__": true, "__values__": ["dry_run task output"'
-                    ', "dry_run task output"]}'
-                ),
-                res,
-                res,
-            )
-
-    @pytest.mark.parametrize("trial", range(5))
-    @pytest.mark.usefixtures("trial")
-    class TestGetWorkflowRunStatus:
-        """
-        Tests that validate .get_workflow_run_status().
-        """
-
-        def test_status_right_after_start(self, runtime: _dag.RayRuntime):
-            """
-            Verifies that we report status correctly before workflow ends.
-            It's difficult to synchronize, so there's a bunch of ifs in this
-            test.
-            """
-            # Given
-            wf_def = _example_wfs.greet_wf.model
-            run_id = runtime.create_workflow_run(wf_def, None, False)
-
-            # When
-            run = runtime.get_workflow_run_status(run_id)
-
-            # Then
-            assert run.id == run_id
-            assert run.status.state in {State.WAITING, State.RUNNING, State.SUCCEEDED}
-            assert run.status.start_time is not None
-            assert run.workflow_def == wf_def
-
-            if run.status.state == State.WAITING:
-                assert run.status.start_time is None
-                assert run.status.end_time is None
-            elif run.status.state == State.RUNNING:
-                assert run.status.start_time is not None
-                assert run.status.end_time is None
-            elif run.status.state == State.SUCCEEDED:
-                assert run.status.start_time is not None
-                assert run.status.end_time is not None
-
-            for task_run in run.task_runs:
-                assert task_run.invocation_id in wf_def.task_invocations.keys()
-
-                status = task_run.status
-                assert status.state in {State.WAITING, State.RUNNING, State.SUCCEEDED}
-
-                if status.state == State.WAITING:
-                    assert status.start_time is None
-                    assert status.end_time is None
-                elif status.state == State.RUNNING:
-                    assert status.start_time is not None
-                    assert status.end_time is None
-                elif status.state == State.SUCCEEDED:
-                    assert status.start_time is not None
-                    assert status.end_time is not None
-
-        def test_status_after_awaiting(self, runtime: _dag.RayRuntime):
-            """
-            Verifies that we report status correctly when all tasks are
-            completed.
-
-            Note: Ray sometimes suffers from race conditions between blocking
-            on the output, and returning workflow status "completed". This is
-            why this test is repeated a couple of times.
-            """
-            # Given
-            wf_def = _example_wfs.greet_wf.model
-            run_id = runtime.create_workflow_run(wf_def, None, False)
-
-            # Block until wf completes
-            _wait_to_finish_wf(run_id, runtime)
-
-            # When
-            run = runtime.get_workflow_run_status(run_id)
-
-            # Then
-            assert run.id == run_id
-            assert (
-                run.status.state == State.SUCCEEDED
-            ), f"Invalid state. Full status: {run.status}. Task runs: {run.task_runs}"
-            assert run.status.start_time is not None
-            assert run.status.end_time is not None
-            assert run.workflow_def == wf_def
-
-            for task_run in run.task_runs:
-                assert task_run.invocation_id in wf_def.task_invocations.keys()
-
-                status = task_run.status
-                assert status.state == State.SUCCEEDED
-
-                assert status.start_time is not None
-                assert status.end_time is not None
-
-                assert status.start_time.tzinfo is not None
-                assert status.end_time.tzinfo is not None
-
-                # On Windows - timer resolution might not detect the change in time
-                # during start and finish of a task. Thus it is >=, not >
-                assert (status.end_time - status.start_time).total_seconds() >= 0
-
-        @freeze_time("9999-01-01")
-        def test_normalizes_end_times(self, runtime: _dag.RayRuntime):
-            # Given
-            wf_def = _example_wfs.serial_wf_with_slow_middle_task.model
-            run_id = runtime.create_workflow_run(wf_def, None, False)
-            runtime.stop_workflow_run(run_id)
-
-            # Block until wf completes
-            _wait_to_finish_wf(run_id, runtime)
-
-            # When
-            run = runtime.get_workflow_run_status(run_id)
-
-            # Then
-            assert run.id == run_id
-            assert (
-                run.status.state == State.TERMINATED
-            ), f"Invalid state. Full status: {run.status}. Task runs: {run.task_runs}"
-            assert run.status.start_time is not None
-            assert run.status.end_time is not None
-            assert run.workflow_def == wf_def
-
-            # The added end times should be 'now'
-            assert run.status.end_time == datetime(9999, 1, 1, tzinfo=timezone.utc)
-            # freeze_time doesn't affect the ray process, so the start times
-            # are real and therefore earlier than the end time (if this is
-            # untrue, hello from the distant past).
-            assert run.status.start_time < run.status.end_time
-
-            for task_run in run.task_runs:
-                status = task_run.status
-
-                # We're only interested in tasks that were terminated while in process.
-                if not (
-                    task_run.status.state == State.TERMINATED
-                    and status.start_time is not None
-                ):
-                    continue
-
-                assert status.end_time is not None
-                assert status.start_time.tzinfo is not None
-                assert status.end_time.tzinfo is not None
-                assert (status.end_time - status.start_time).total_seconds() >= 0
-                assert task_run.status.end_time == datetime(
-                    9999, 1, 1, tzinfo=timezone.utc
-                )
-
-        @pytest.mark.skipif(
-            sys.platform.startswith("win32"), reason="File writing on windows is slow."
-        )
-        def test_handles_ray_environment_setup_error(
-            self, runtime: _dag.RayRuntime, shared_ray_conn
-        ):
-            # Given
-            wf_def = _example_wfs.cause_env_setup_error.model
-            run_id = runtime.create_workflow_run(wf_def, None, False)
-
-            # Block until wf completes
-            _wait_to_finish_wf(run_id, runtime, timeout=100)
-
-            # When
-            run = runtime.get_workflow_run_status(run_id)
-
-            # Then
-            assert (
-                run.status.state == State.FAILED
-            ), f"Invalid state. Full status: {run.status}. Task runs: {run.task_runs}"
-            # Grab the logs so we can debug when the test fails
-            logs = _ray_logs.DirectLogReader(
-                Path(shared_ray_conn._temp_dir)
-            ).get_workflow_logs(wf_run_id=run_id)
-
-            re_pattern = (
-                r"Could not set up runtime environment "
-                r"\('pip\.py:\d* -- Failed to install pip packages'\)\. "
-                r"See environment setup logs for details. "
-                r"`orq wf logs " + re.escape(str(run_id)) + r" --env-setup`"
-            )
-
-            assert re.match(re_pattern, str(run.message)), (
-                f"\n-MESSAGE: {run.message}"
-                f"\n-ENV SETUP OUT:\n{logs.env_setup.out}"
-                f"\n-ENV SETUP ERR:\n{logs.env_setup.err}"
-                f"\n-OTHER:\n{logs}"
-            )
-
-        def test_exception_in_task_stops_execution(self, runtime: _dag.RayRuntime):
-            """
-            The workflow graph:
-
-               [1]
-                │
-                ▼
-               [2] => exception
-                │
-                ▼
-               [3] => won't run
-                │
-                ▼
-            [return]
-            """
-            wf_def = _example_wfs.exception_wf_with_multiple_values.model
-            run_id = runtime.create_workflow_run(wf_def, None, False)
-
-            _wait_to_finish_wf(run_id, runtime)
-
-            wf_run = runtime.get_workflow_run_status(run_id)
-
-            assert wf_run.status.state == State.FAILED
-            assert _count_task_runs(wf_run, State.FAILED) == 1
-            assert _count_task_runs(wf_run, State.SUCCEEDED) == 1
-            assert _count_task_runs(wf_run, State.WAITING) == 1
-
-    class TestListWorkflowRuns:
-        """
-        Tests that validate .list_workflow_runs().
-        """
-
-        def test_two_runs(self, runtime: _dag.RayRuntime):
-            # Given
-            wf_def = _example_wfs.greet_wf.model
-            run_id1 = runtime.create_workflow_run(wf_def, None, False)
-            run_id2 = runtime.create_workflow_run(wf_def, None, False)
-
-            # When
-            wf_runs = runtime.list_workflow_runs()
-
-            # Then
-            wf_run_ids = {run.id for run in wf_runs}
-            assert run_id1 in wf_run_ids
-            assert run_id2 in wf_run_ids
-            assert all(isinstance(run, WorkflowRun) for run in wf_runs)
-
-    class TestStopWorkflow:
-        """
-        Tests that validate .stop_workflow_run()
-        """
-
-        def test_on_running_workflow(self, runtime: _dag.RayRuntime, tmp_path):
-            """
-            Workflow graph in the scenario under test:
-              [ ]  => in progress
-               │
-               │
-               ▼
-            """
-
-            wf = _example_wfs.infinite_workflow().model
-
-            wf_run_id = runtime.create_workflow_run(wf, None, False)
-            wf_run = runtime.get_workflow_run_status(wf_run_id)
-            assert wf_run.status.state == State.RUNNING
-
-            runtime.stop_workflow_run(wf_run_id)
-
-            # ray should cancel workflow synchronously, but just in case it doesn't
-            # let's give a workflow some time to change its state
-            timeout = time.time() + 30  # now + 30 seconds
-            while True:
-                if time.time() > timeout:
-                    assert False, "timeout while waiting for workflow termination"
-                wf_run = runtime.get_workflow_run_status(wf_run_id)
-                if wf_run.status.state != State.RUNNING:
-                    assert wf_run.status.state == State.TERMINATED
-                    break
-
-        @pytest.mark.filterwarnings("ignore::pytest.PytestUnraisableExceptionWarning")
-        def test_on_finished_workflow(self, runtime: _dag.RayRuntime, tmp_path):
-            wf = _example_wfs.multioutput_task_wf.model
-            wf_run_id = runtime.create_workflow_run(wf, None, False)
-            _wait_to_finish_wf(wf_run_id, runtime)
-            # ensure wf has finished
-            status = runtime.get_workflow_run_status(wf_run_id)
-            assert status.status.state == State.SUCCEEDED
-
-            runtime.stop_workflow_run(wf_run_id)
-            status = runtime.get_workflow_run_status(wf_run_id)
-            # cancel changes the state to terminated
-            assert status.status.state == State.TERMINATED
-
-    @pytest.mark.filterwarnings("ignore::pytest.PytestUnraisableExceptionWarning")
-    class TestGetWorkflowRunOutputsNonBlocking:
-        """
-        Tests that validate get_workflow_run_outputs_non_blocking
-        """
-
-        def test_happy_path(self, runtime: _dag.RayRuntime):
-            wf_def = _example_wfs.greet_wf.model
-            run_id = runtime.create_workflow_run(wf_def, None, False)
-
-            _wait_to_finish_wf(run_id, runtime)
-
-            outputs = runtime.get_workflow_run_outputs_non_blocking(run_id)
-
-            assert outputs == (
-                JSONResult(value='"yooooo emiliano from zapata computing"'),
-            )
-
-        def test_failed_workflow(self, runtime: _dag.RayRuntime):
-            wf_def = _example_wfs.exception_wf_with_multiple_values().model
-            run_id = runtime.create_workflow_run(wf_def, None, False)
-
-            _wait_to_finish_wf(run_id, runtime)
-
-            with pytest.raises(exceptions.WorkflowRunNotSucceeded):
-                runtime.get_workflow_run_outputs_non_blocking(run_id)
-
-        def test_in_progress_workflow(self, runtime: _dag.RayRuntime, tmp_path):
-            """
-            Workflow graph in the scenario under test:
-              [ ]  => in progress
-               │
-               │
-               ▼
-              [ ]  => waiting
-               │
-               │
-               ▼
-            """
-            triggers = [_ipc.TriggerServer() for _ in range(2)]
-
-            wf = _example_wfs.serial_wf_with_file_triggers(
-                [trigger.port for trigger in triggers], task_timeout=5.0
-            ).model
-            run_id = runtime.create_workflow_run(wf, None, False)
-
-            assert runtime.get_workflow_run_status(run_id).status.state == State.RUNNING
-            with pytest.raises(exceptions.WorkflowRunNotSucceeded):
-                runtime.get_workflow_run_outputs_non_blocking(run_id)
-
-            for trigger in triggers:
-                trigger.trigger()
-                trigger.close()
-
-    class TestGetAvailableOutputs:
-        """
-        Tests that validate .get_available_outputs().
-        """
-
-        def test_after_a_failed_task(self, runtime: _dag.RayRuntime):
-            """
-            The workflow graph:
-
-               [1]
-                │
-                ▼
-               [2] => exception
-                │
-                ▼
-               [3] => won't run
-                │
-                ▼
-            [return]
-            """
-            wf_def = _example_wfs.exception_wf_with_multiple_values().model
-            run_id = runtime.create_workflow_run(wf_def, None, False)
-
-            _wait_to_finish_wf(run_id, runtime)
-
-            wf_run = runtime.get_workflow_run_status(run_id)
-            inv_artifacts = runtime.get_available_outputs(run_id)
-            if wf_run.status.state != State.FAILED:
-                pytest.fail(
-                    "The workflow was supposed to fail, but it didn't. "
-                    f"Wf run: {wf_run}"
-                )
-
-            # Expect only one finished task invocation.
-            assert len(inv_artifacts) == 1
-
-            # The outputs dict has an entry for each task invocation. The value is
-            # whatever the task returned; in this example it's a single int.
-            task_output = list(inv_artifacts.values())[0]
-            assert task_output == JSONResult(value="58")
-
-        @pytest.mark.filterwarnings("ignore::pytest.PytestUnraisableExceptionWarning")
-        def test_after_one_task_finishes(self, runtime: _dag.RayRuntime, tmp_path):
-            """
-            Workflow graph in the scenario under test:
-              [ ]  => finished
-               │
-               │
-               ▼
-              [ ]  => in progress
-               │
-               │
-               ▼
-              [ ]  => waiting
-               │
-               │
-               ▼
-            """
-            triggers = [_ipc.TriggerServer() for _ in range(3)]
-
-            wf = _example_wfs.serial_wf_with_file_triggers(
-                [trigger.port for trigger in triggers], task_timeout=10.0
-            ).model
-            wf_run_id = runtime.create_workflow_run(wf, None, False)
-
-            triggers[0].trigger()
-            # Await completion of the first task
-            loop_start = time.time()
-            while True:
-                wf_run = runtime.get_workflow_run_status(wf_run_id)
-                if wf_run.status.state not in {State.RUNNING, State.SUCCEEDED}:
-                    pytest.fail(f"Unexpected wf run state: {wf_run}")
-
-                succeeded_runs = [
-                    run
-                    for run in wf_run.task_runs
-                    if run.status.state == State.SUCCEEDED
-                ]
-                if len(succeeded_runs) >= 1:
-                    break
-
-                if time.time() - loop_start > 20.0:
-                    pytest.fail(
-                        f"Timeout when awaiting for workflow finish. Full run: {wf_run}"
-                    )
-
-                time.sleep(0.2)
-
-            # When
-            outputs_dict = runtime.get_available_outputs(wf_run_id)
-
-            # Then
-            assert len(outputs_dict) == 1
-            assert len(succeeded_runs) == 1
-
-            # let the workers complete the workflow
-            triggers[1].trigger()
-            triggers[2].trigger()
-
-            for trigger in triggers:
-                trigger.close()
-
-    class TestGetOutput:
-        """
-        Tests that validate .get_output().
-        """
-
-        def test_after_a_failed_task(self, runtime: _dag.RayRuntime):
-            """
-            The workflow graph:
-
-               [1]
-                │
-                ▼
-               [2] => exception
-                │
-                ▼
-               [3] => won't run
-                │
-                ▼
-            [return]
-            """
-            wf_def = _example_wfs.exception_wf_with_multiple_values().model
-            run_id = runtime.create_workflow_run(wf_def, None, False)
-
-            _wait_to_finish_wf(run_id, runtime)
-
-            wf_run = runtime.get_workflow_run_status(run_id)
-            invocation_ids = [task.invocation_id for task in wf_run.task_runs]
-
-            if wf_run.status.state != State.FAILED:
-                pytest.fail(
-                    "The workflow was supposed to fail, but it didn't. "
-                    f"Wf run: {wf_run}"
-                )
-
-            assert runtime.get_output(run_id, invocation_ids[2]) == JSONResult(
-                value="58"
-            )
-            with pytest.raises(exceptions.NotFoundError):
-                assert not runtime.get_output(run_id, invocation_ids[1])
-            with pytest.raises(exceptions.NotFoundError):
-                assert not runtime.get_output(run_id, invocation_ids[0])
-
-        @pytest.mark.filterwarnings("ignore::pytest.PytestUnraisableExceptionWarning")
-        def test_after_one_task_finishes(self, runtime: _dag.RayRuntime, tmp_path):
-            """
-            Workflow graph in the scenario under test:
-              [ ]  => finished
-               │
-               │
-               ▼
-              [ ]  => in progress
-               │
-               │
-               ▼
-              [ ]  => waiting
-               │
-               │
-               ▼
-            """
-            triggers = [_ipc.TriggerServer() for _ in range(3)]
-
-            wf = _example_wfs.serial_wf_with_file_triggers(
-                [trigger.port for trigger in triggers], task_timeout=10.0
-            ).model
-            wf_run_id = runtime.create_workflow_run(wf, None, False)
-
-            triggers[0].trigger()
-            # Await completion of the first task
-            loop_start = time.time()
-            while True:
-                wf_run = runtime.get_workflow_run_status(wf_run_id)
-                if wf_run.status.state not in {State.RUNNING, State.SUCCEEDED}:
-                    pytest.fail(f"Unexpected wf run state: {wf_run}")
-
-                succeeded_runs = [
-                    run
-                    for run in wf_run.task_runs
-                    if run.status.state == State.SUCCEEDED
-                ]
-                if len(succeeded_runs) >= 1:
-                    break
-
-                if time.time() - loop_start > 20.0:
-                    pytest.fail(
-                        f"Timeout when awaiting for workflow finish. Full run: {wf_run}"
-                    )
-
-                time.sleep(0.2)
-            invocation_ids = [task.invocation_id for task in wf_run.task_runs]
-
-            assert runtime.get_output(wf_run_id, invocation_ids[2]) == JSONResult(
-                value="58"
-            )
-            with pytest.raises(exceptions.NotFoundError):
-                assert not runtime.get_output(wf_run_id, invocation_ids[1])
-            with pytest.raises(exceptions.NotFoundError):
-                assert not runtime.get_output(wf_run_id, invocation_ids[0])
-
-            # let the workers complete the workflow
-            triggers[1].trigger()
-            triggers[2].trigger()
-
-            for trigger in triggers:
-                trigger.close()
-
-        @pytest.mark.filterwarnings("ignore::pytest.PytestUnraisableExceptionWarning")
-        def test_multioutput_task(self, runtime: _dag.RayRuntime, tmp_path):
-            wf = _example_wfs.multioutput_task_wf().model
-            wf_run_id = runtime.create_workflow_run(wf, None, False)
-
-            _wait_to_finish_wf(wf_run_id, runtime)
-
-            wf_run = runtime.get_workflow_run_status(wf_run_id)
-
-            invocation_ids = [task.invocation_id for task in wf_run.task_runs]
-            expected_result = (
-                '{"__tuple__": true, "__values__": ["Zapata", "Computing"]}'
-            )
-
-            artifacts = [
-                runtime.get_output(wf_run_id, inv_id) for inv_id in invocation_ids
-            ]
-
-            assert len(artifacts) == 4
-            assert all(
-                [
-                    artifact == JSONResult(value=expected_result)
-                    for artifact in artifacts
-                ]
-            )
-
-            assert all(
-                [
-                    deserialize(artifact) == ("Zapata", "Computing")
-                    for artifact in artifacts
-                ]
-            )
 
 
 @pytest.mark.slow
@@ -1584,28 +919,6 @@ class TestMakeDag:
     def wf_run_id(self):
         return "mocked_wf_run_id"
 
-    def test_import_pip_string_resolved_once_per_import(
-        self, monkeypatch: pytest.MonkeyPatch, client: Mock, wf_run_id: str
-    ):
-        pip_string = create_autospec(_build_workflow._pip_string)
-        pip_string.return_value = ["mocked"]
-        monkeypatch.setattr(_build_workflow, "_pip_string", pip_string)
-        imps = [
-            sdk.GithubImport(
-                "zapata-engineering/orquestra-sdk",
-                personal_access_token=sdk.Secret(
-                    "mock-secret", config_name="mock-config", workspace_id="mock-ws"
-                ),
-            ),
-            sdk.PythonImports("numpy", "polars"),
-            sdk.InlineImport(),
-        ]
-        workflow_def = make_workflow_with_dependencies(imps, n_tasks=100).model
-        _ = _build_workflow.make_ray_dag(client, workflow_def, wf_run_id, False)
-        assert pip_string.mock_calls == [
-            call(imp) for imp in workflow_def.imports.values()
-        ]
-
     class TestResourcesInMakeDag:
         @pytest.mark.parametrize(
             "resources, expected, types",
@@ -1759,177 +1072,156 @@ class TestMakeDag:
                     **expected_kwargs,
                 )
 
+@pytest.mark.slow
+class TestRayRuntimeMethods:
+    def test_in_progress_workflow(self, runtime: _dag.RayRuntime, tmp_path):
+        """
+        Workflow graph in the scenario under test:
+          [ ]  => in progress
+           │
+           │
+           ▼
+          [ ]  => waiting
+           │
+           │
+           ▼
+        """
+        triggers = [_ipc.TriggerServer() for _ in range(2)]
 
-@pytest.mark.parametrize(
-    "installed_sdk_version, expected_sdk_dependency",
-    [
-        ("0.57.1.dev3+g3ef9f57.d20231003", None),
-        ("1.2.3", "1.2.3"),
-        ("0.1.dev1+g25df81e", None),
-    ],
-)
-class TestHandlingSDKVersions:
-    """``_import_pip_env`` handles adding the current SDK version as a dependency.
+        wf = _example_wfs.serial_wf_with_file_triggers(
+            [trigger.port for trigger in triggers], task_timeout=5.0
+        ).model
+        run_id = runtime.create_workflow_run(wf, None, False)
 
-    Note that these tests don't mock serde - we're interested in whether the correct
-    package list gets constructed so we can't just return 'mocked' for imports.
-    """
+        assert runtime.get_workflow_run_status(run_id).status.state == State.RUNNING
+        with pytest.raises(exceptions.WorkflowRunNotSucceeded):
+            runtime.get_workflow_run_outputs_non_blocking(run_id)
 
-    @staticmethod
-    def test_with_no_dependencies(
-        installed_sdk_version: str,
-        expected_sdk_dependency: str,
-        monkeypatch: pytest.MonkeyPatch,
-    ):
-        # Given
-        monkeypatch.setattr(
-            _build_workflow,
-            "get_installed_version",
-            Mock(return_value=installed_sdk_version),
-        )
-        wf = make_workflow_with_dependencies([]).model
-        task_inv = [inv for inv in iter_invocations_topologically(wf)][0]
-        imports = {
-            id: _build_workflow._pip_string(imp) for id, imp in wf.imports.items()
-        }
+        for trigger in triggers:
+            trigger.trigger()
+            trigger.close()
 
-        # When
-        pip = _build_workflow._import_pip_env(task_inv, wf, imports)
 
-        # Then
-        if expected_sdk_dependency:
-            assert pip == [f"orquestra-sdk=={expected_sdk_dependency}"]
-        else:
-            assert pip == []
+    @pytest.mark.filterwarnings("ignore::pytest.PytestUnraisableExceptionWarning")
+    def test_after_one_task_finishes(self, runtime: _dag.RayRuntime, tmp_path):
+        """
+        Workflow graph in the scenario under test:
+          [ ]  => finished
+           │
+           │
+           ▼
+          [ ]  => in progress
+           │
+           │
+           ▼
+          [ ]  => waiting
+           │
+           │
+           ▼
+        """
+        triggers = [_ipc.TriggerServer() for _ in range(3)]
 
-    @staticmethod
-    @pytest.mark.parametrize(
-        "python_imports",
-        [
-            ["MarkupSafe==1.0.0"],
-            ["MarkupSafe==1.0.0", "Jinja2==2.7.2"],
-        ],
-    )
-    def test_with_multiple_dependencies(
-        python_imports: t.List[str],
-        installed_sdk_version: str,
-        expected_sdk_dependency: str,
-        monkeypatch: pytest.MonkeyPatch,
-    ):
-        # Given
-        monkeypatch.setattr(
-            _build_workflow,
-            "get_installed_version",
-            Mock(return_value=installed_sdk_version),
-        )
-        wf = make_workflow_with_dependencies([sdk.PythonImports(*python_imports)]).model
-        task_inv = [inv for inv in iter_invocations_topologically(wf)][0]
-        imports = {
-            id: _build_workflow._pip_string(imp) for id, imp in wf.imports.items()
-        }
+        wf = _example_wfs.serial_wf_with_file_triggers(
+            [trigger.port for trigger in triggers], task_timeout=10.0
+        ).model
+        wf_run_id = runtime.create_workflow_run(wf, None, False)
 
-        # When
-        pip = _build_workflow._import_pip_env(task_inv, wf, imports)
+        triggers[0].trigger()
+        # Await completion of the first task
+        loop_start = time.time()
+        while True:
+            wf_run = runtime.get_workflow_run_status(wf_run_id)
+            if wf_run.status.state not in {State.RUNNING, State.SUCCEEDED}:
+                pytest.fail(f"Unexpected wf run state: {wf_run}")
 
-        # Then
-        if expected_sdk_dependency:
-            assert sorted(pip) == sorted(
-                python_imports + [f"orquestra-sdk=={expected_sdk_dependency}"]
-            )
-        else:
-            assert sorted(pip) == sorted(python_imports)
+            succeeded_runs = [
+                run
+                for run in wf_run.task_runs
+                if run.status.state == State.SUCCEEDED
+            ]
+            if len(succeeded_runs) >= 1:
+                break
 
-    @pytest.mark.parametrize(
-        "sdk_import",
-        [
-            "orquestra-sdk",
-            "orquestra-sdk>=1.2.3",
-            "orquestra-sdk<=1.2.3",
-            "orquestra-sdk~=1.2.3",
-            "orquestra-sdk==1.2.3",
-            "orquestra-sdk!=1.2.3",
-            "orquestra-sdk===1.2.3",
-            "orquestra-sdk >= 1.2.3",
-            "orquestra-sdk <= 1.2.3",
-            "orquestra-sdk ~= 1.2.3",
-            "orquestra-sdk == 1.2.3",
-            "orquestra-sdk != 1.2.3",
-            "orquestra-sdk === 1.2.3",
-        ],
-    )
-    @pytest.mark.parametrize(
-        "python_imports",
-        [
-            [],
-            ["MarkupSafe==1.0.0"],
-            ["MarkupSafe==1.0.0", "Jinja2==2.7.2"],
-        ],
-    )
-    class TestHandlesSDKDependency:
-        @staticmethod
-        @pytest.mark.filterwarnings("ignore:The definition for task ")
-        def test_replaces_declared_sdk_dependency(
-            sdk_import,
-            python_imports,
-            installed_sdk_version: str,
-            expected_sdk_dependency: str,
-            monkeypatch: pytest.MonkeyPatch,
-        ):
-            # Given
-            monkeypatch.setattr(
-                _build_workflow,
-                "get_installed_version",
-                Mock(return_value=installed_sdk_version),
-            )
-            wf = make_workflow_with_dependencies(
-                [sdk.PythonImports(*python_imports + [sdk_import])]
-            ).model
-            task_inv = [inv for inv in iter_invocations_topologically(wf)][0]
-            imports = {
-                id: _build_workflow._pip_string(imp) for id, imp in wf.imports.items()
-            }
-
-            # When
-            pip = _build_workflow._import_pip_env(task_inv, wf, imports)
-
-            # Then
-            if expected_sdk_dependency:
-                assert sorted(pip) == sorted(
-                    python_imports + [f"orquestra-sdk=={expected_sdk_dependency}"]
+            if time.time() - loop_start > 20.0:
+                pytest.fail(
+                    f"Timeout when awaiting for workflow finish. Full run: {wf_run}"
                 )
-            else:
-                assert sorted(pip) == sorted(python_imports)
 
-        @staticmethod
-        @pytest.mark.filterwarnings("error")
-        def test_warns_user_that_declared_sdk_dependency_is_ignored(
-            sdk_import,
-            python_imports,
-            installed_sdk_version: str,
-            expected_sdk_dependency: str,
-            monkeypatch: pytest.MonkeyPatch,
-        ):
-            # Given
-            monkeypatch.setattr(
-                _build_workflow,
-                "get_installed_version",
-                Mock(return_value=installed_sdk_version),
-            )
-            wf = make_workflow_with_dependencies(
-                [sdk.PythonImports(*python_imports + [sdk_import])]
-            ).model
-            task_inv = [inv for inv in iter_invocations_topologically(wf)][0]
-            imports = {
-                id: _build_workflow._pip_string(imp) for id, imp in wf.imports.items()
-            }
+            time.sleep(0.2)
 
-            # When
-            with pytest.raises(OrquestraSDKVersionMismatchWarning) as e:
-                _ = _build_workflow._import_pip_env(task_inv, wf, imports)
+        # When
+        outputs_dict = runtime.get_available_outputs(wf_run_id)
 
-            # Then
-            warning: str = e.exconly().strip()
-            assert re.match(
-                r"^orquestra\.workflow_shared\.exceptions\.OrquestraSDKVersionMismatchWarning: The definition for task `task-hello-orquestra-.*` declares `orquestra-sdk(?P<dependency>.*)` as a dependency. The current SDK version (\((?P<installed>.*)\) )?is automatically installed in task environments. The specified dependency will be ignored.$",  # noqa: E501
-                warning,
-            ), warning
+        # Then
+        assert len(outputs_dict) == 1
+        assert len(succeeded_runs) == 1
+
+        # let the workers complete the workflow
+        triggers[1].trigger()
+        triggers[2].trigger()
+
+        for trigger in triggers:
+            trigger.close()
+
+    @pytest.mark.filterwarnings("ignore::pytest.PytestUnraisableExceptionWarning")
+    def test_after_one_task_finishes(self, runtime: _dag.RayRuntime, tmp_path):
+        """
+        Workflow graph in the scenario under test:
+          [ ]  => finished
+           │
+           │
+           ▼
+          [ ]  => in progress
+           │
+           │
+           ▼
+          [ ]  => waiting
+           │
+           │
+           ▼
+        """
+        triggers = [_ipc.TriggerServer() for _ in range(3)]
+
+        wf = _example_wfs.serial_wf_with_file_triggers(
+            [trigger.port for trigger in triggers], task_timeout=10.0
+        ).model
+        wf_run_id = runtime.create_workflow_run(wf, None, False)
+
+        triggers[0].trigger()
+        # Await completion of the first task
+        loop_start = time.time()
+        while True:
+            wf_run = runtime.get_workflow_run_status(wf_run_id)
+            if wf_run.status.state not in {State.RUNNING, State.SUCCEEDED}:
+                pytest.fail(f"Unexpected wf run state: {wf_run}")
+
+            succeeded_runs = [
+                run
+                for run in wf_run.task_runs
+                if run.status.state == State.SUCCEEDED
+            ]
+            if len(succeeded_runs) >= 1:
+                break
+
+            if time.time() - loop_start > 20.0:
+                pytest.fail(
+                    f"Timeout when awaiting for workflow finish. Full run: {wf_run}"
+                )
+
+            time.sleep(0.2)
+        invocation_ids = [task.invocation_id for task in wf_run.task_runs]
+
+        assert runtime.get_output(wf_run_id, invocation_ids[2]) == JSONResult(
+            value="58"
+        )
+        with pytest.raises(exceptions.NotFoundError):
+            assert not runtime.get_output(wf_run_id, invocation_ids[1])
+        with pytest.raises(exceptions.NotFoundError):
+            assert not runtime.get_output(wf_run_id, invocation_ids[0])
+
+        # let the workers complete the workflow
+        triggers[1].trigger()
+        triggers[2].trigger()
+
+        for trigger in triggers:
+            trigger.close()
