+++ projects/orquestra-workflow-runtime/tests/__init__.py
@@ -1,0 +1,5 @@
+################################################################################
+# © Copyright 2024 Zapata Computing Inc.
+################################################################################
+"""Required by pytest to figure out test modules properly. We use that for
+testing utilities and resolving test workflows & tasks properly."""
