+++ projects/orquestra-workflow-runtime/tests/data/cause_env_setup_error.json
@@ -1,0 +1,1 @@
+"{\"name\":\"cause_env_setup_error\",\"fn_ref\":{\"module\":\"__main__\",\"function_name\":\"cause_env_setup_error\",\"file_path\":\"generate_workflow_defs.py\",\"line_number\":244,\"type\":\"MODULE_FUNCTION_REF\"},\"imports\":{\"inline-import-11\":{\"id\":\"inline-import-11\",\"type\":\"INLINE_IMPORT\"},\"python-import-b1a1ddb43d\":{\"id\":\"python-import-b1a1ddb43d\",\"packages\":[{\"name\":\"MarkupSafe\",\"extras\":[],\"version_constraints\":[\"==1.0.0\"],\"environment_markers\":\"\"},{\"name\":\"Jinja2\",\"extras\":[],\"version_constraints\":[\"==2.7.2\"],\"environment_markers\":\"\"}],\"pip_options\":[],\"type\":\"PYTHON_IMPORT\"}},\"tasks\":{\"task-cause-env-setup-error-task-8cb7192be3\":{\"id\":\"task-cause-env-setup-error-task-8cb7192be3\",\"fn_ref\":{\"function_name\":\"cause_env_setup_error_task\",\"encoded_function\":[\"gASVuAQAAAAAAACMF2Nsb3VkcGlja2xlLmNsb3VkcGlja2xllIwOX21ha2VfZnVuY3Rpb26Uk5Qo\\naACMDV9idWlsdGluX3R5cGWUk5SMCENvZGVUeXBllIWUUpQoSwBLAEsASwBLAUsTQwaXAGQBUwCU\\nToiGlCkpjHwvVXNlcnMvc2ViYXN0aWFubW9yYXdpZWMvZ2l0L29ycXVlc3RyYS13b3JrZmxvdy1z\\nZGsvcHJvamVjdHMvb3JxdWVzdHJhLXdvcmtmbG93LXJ1bnRpbWUvdGVzdHMvZGF0YS9nZW5lcmF0\\nZV93b3JrZmxvd19kZWZzLnB5lIwaY2F1c2VfZW52X3NldHVwX2Vycm9yX3Rhc2uUjEJnZW5lcmF0\\nZV9jYXVzZV9lbnZfc2V0dXBfZXJyb3IuPGxvY2Fscz4uY2F1c2VfZW52X3NldHVwX2Vycm9yX3Rh\\nc2uUS+1DCYAA8AoAEBSIdJRDAJQpKXSUUpR9lCiMC19fcGFja2FnZV9flE6MCF9fbmFtZV9flIwI\\nX19tYWluX1+UjAhfX2ZpbGVfX5RoCnVOTk50lFKUjBxjbG91ZHBpY2tsZS5jbG91ZHBpY2tsZV9m\\nYXN0lIwSX2Z1bmN0aW9uX3NldHN0YXRllJOUaBd9lCiMF19UYXNrRGVmX19zZGtfdGFza19ib2R5\\nlGgXjAdfZm5fcmVmlIwgb3JxdWVzdHJhLnNkay5fY2xpZW50Ll9iYXNlLl9kc2yUjBFJbmxpbmVG\\ndW5jdGlvblJlZpSTlGgLaBeGlIGUjAhfZm5fbmFtZZRoC4wQX291dHB1dF9tZXRhZGF0YZRoHowS\\nVGFza091dHB1dE1ldGFkYXRhlJOUiUsBhpSBlIwLX3BhcmFtZXRlcnOUjAtjb2xsZWN0aW9uc5SM\\nC09yZGVyZWREaWN0lJOUKVKUjApfcmVzb3VyY2VzlGgejAlSZXNvdXJjZXOUk5QoTk5OTk50lIGU\\njA1fY3VzdG9tX2ltYWdllE6MDF9jdXN0b21fbmFtZZROjBNfZGVwZW5kZW5jeV9pbXBvcnRzlGge\\njA1QeXRob25JbXBvcnRzlJOUKYGUfZQojAVfZmlsZZROjAlfcGFja2FnZXOUjBFNYXJrdXBTYWZl\\nPT0xLjAuMJSMDUppbmphMj09Mi43LjKUhpR1YoWUjB9fdXNlX2RlZmF1bHRfZGVwZW5kZW5jeV9p\\nbXBvcnRzlImMDl9zb3VyY2VfaW1wb3J0lGgejAxJbmxpbmVJbXBvcnSUk5QpgZSMGl91c2VfZGVm\\nYXVsdF9zb3VyY2VfaW1wb3J0lImMDF9tYXhfcmV0cmllc5ROjAlfZW52X3ZhcnOUTnV9lChoE2gL\\njAxfX3F1YWxuYW1lX1+UaAyMD19fYW5ub3RhdGlvbnNfX5R9lIwGcmV0dXJulIwIYnVpbHRpbnOU\\njARib29slJOUc4wOX19rd2RlZmF1bHRzX1+UTowMX19kZWZhdWx0c19flE6MCl9fbW9kdWxlX1+U\\naBSMB19fZG9jX1+UTowLX19jbG9zdXJlX1+UTowXX2Nsb3VkcGlja2xlX3N1Ym1vZHVsZXOUXZSM\\nC19fZ2xvYmFsc19flH2UdYaUhlIwLg==\\n\"],\"type\":\"INLINE_FUNCTION_REF\"},\"parameters\":[],\"output_metadata\":{\"is_subscriptable\":false,\"n_outputs\":1},\"source_import_id\":\"inline-import-11\",\"dependency_import_ids\":[\"python-import-b1a1ddb43d\"],\"resources\":null,\"max_retries\":null,\"custom_image\":null}},\"artifact_nodes\":{\"artifact-0-cause-env-setup-error-task\":{\"id\":\"artifact-0-cause-env-setup-error-task\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":null}},\"constant_nodes\":{},\"secret_nodes\":{},\"task_invocations\":{\"invocation-0-task-cause-env-setup-error-task\":{\"id\":\"invocation-0-task-cause-env-setup-error-task\",\"task_id\":\"task-cause-env-setup-error-task-8cb7192be3\",\"args_ids\":[],\"kwargs_ids\":{},\"output_ids\":[\"artifact-0-cause-env-setup-error-task\"],\"resources\":null,\"custom_image\":null,\"env_vars\":null}},\"output_ids\":[\"artifact-0-cause-env-setup-error-task\"],\"data_aggregation\":null,\"metadata\":{\"sdk_version\":{\"original\":\"0.64.0\",\"major\":0,\"minor\":64,\"patch\":0,\"is_prerelease\":false},\"python_version\":{\"original\":\"3.11.6 (main, Nov  6 2023, 13:24:14) [Clang 15.0.0 (clang-1500.0.40.1)]\",\"major\":3,\"minor\":11,\"patch\":6,\"is_prerelease\":false}},\"resources\":null}"
