+++ projects/orquestra-workflow-runtime/tests/data/exception_wf_with_multiple_values.json
@@ -1,0 +1,1 @@
+"{\"name\":\"exception_wf_with_multiple_values\",\"fn_ref\":{\"module\":\"__main__\",\"function_name\":\"exception_wf_with_multiple_values\",\"file_path\":\"generate_workflow_defs.py\",\"line_number\":141,\"type\":\"MODULE_FUNCTION_REF\"},\"imports\":{\"inline-import-6\":{\"id\":\"inline-import-6\",\"type\":\"INLINE_IMPORT\"}},\"tasks\":{\"task-add-e573a4e53a\":{\"id\":\"task-add-e573a4e53a\",\"fn_ref\":{\"function_name\":\"add\",\"encoded_function\":[\"gASVUgQAAAAAAACMF2Nsb3VkcGlja2xlLmNsb3VkcGlja2xllIwOX21ha2VfZnVuY3Rpb26Uk5Qo\\naACMDV9idWlsdGluX3R5cGWUk5SMCENvZGVUeXBllIWUUpQoSwJLAEsASwJLAksDQwyXAHwAfAF6\\nAAAAUwCUToWUKYwBYZSMAWKUhpSMfC9Vc2Vycy9zZWJhc3RpYW5tb3Jhd2llYy9naXQvb3JxdWVz\\ndHJhLXdvcmtmbG93LXNkay9wcm9qZWN0cy9vcnF1ZXN0cmEtd29ya2Zsb3ctcnVudGltZS90ZXN0\\ncy9kYXRhL2dlbmVyYXRlX3dvcmtmbG93X2RlZnMucHmUjANhZGSUaA5LQUMLgADgCwyIcYk1gEyU\\nQwCUKSl0lFKUfZQojAtfX3BhY2thZ2VfX5ROjAhfX25hbWVfX5SMCF9fbWFpbl9flIwIX19maWxl\\nX1+UaA11Tk5OdJRSlIwcY2xvdWRwaWNrbGUuY2xvdWRwaWNrbGVfZmFzdJSMEl9mdW5jdGlvbl9z\\nZXRzdGF0ZZSTlGgZfZQojBdfVGFza0RlZl9fc2RrX3Rhc2tfYm9keZRoGYwHX2ZuX3JlZpSMIG9y\\ncXVlc3RyYS5zZGsuX2NsaWVudC5fYmFzZS5fZHNslIwRSW5saW5lRnVuY3Rpb25SZWaUk5RoDmgZ\\nhpSBlIwIX2ZuX25hbWWUaA6MEF9vdXRwdXRfbWV0YWRhdGGUaCCMElRhc2tPdXRwdXRNZXRhZGF0\\nYZSTlIlLAYaUgZSMC19wYXJhbWV0ZXJzlIwLY29sbGVjdGlvbnOUjAtPcmRlcmVkRGljdJSTlClS\\nlChoCmggjA1UYXNrUGFyYW1ldGVylJOUaApoIIwNUGFyYW1ldGVyS2luZJSTlIwVUE9TSVRJT05B\\nTF9PUl9LRVlXT1JElIWUUpSGlIGUaAtoMWgLaDaGlIGUdYwKX3Jlc291cmNlc5RoIIwJUmVzb3Vy\\nY2VzlJOUKE5OTk5OdJSBlIwNX2N1c3RvbV9pbWFnZZROjAxfY3VzdG9tX25hbWWUTowTX2RlcGVu\\nZGVuY3lfaW1wb3J0c5ROjB9fdXNlX2RlZmF1bHRfZGVwZW5kZW5jeV9pbXBvcnRzlIiMDl9zb3Vy\\nY2VfaW1wb3J0lGggjAxJbmxpbmVJbXBvcnSUk5QpgZSMGl91c2VfZGVmYXVsdF9zb3VyY2VfaW1w\\nb3J0lIiMDF9tYXhfcmV0cmllc5ROjAlfZW52X3ZhcnOUTnV9lChoFWgOjAxfX3F1YWxuYW1lX1+U\\naA6MD19fYW5ub3RhdGlvbnNfX5R9lIwOX19rd2RlZmF1bHRzX1+UTowMX19kZWZhdWx0c19flE6M\\nCl9fbW9kdWxlX1+UaBaMB19fZG9jX1+UTowLX19jbG9zdXJlX1+UTowXX2Nsb3VkcGlja2xlX3N1\\nYm1vZHVsZXOUXZSMC19fZ2xvYmFsc19flH2UdYaUhlIwLg==\\n\"],\"type\":\"INLINE_FUNCTION_REF\"},\"parameters\":[{\"name\":\"a\",\"kind\":\"POSITIONAL_OR_KEYWORD\"},{\"name\":\"b\",\"kind\":\"POSITIONAL_OR_KEYWORD\"}],\"output_metadata\":{\"is_subscriptable\":false,\"n_outputs\":1},\"source_import_id\":\"inline-import-6\",\"dependency_import_ids\":null,\"resources\":null,\"max_retries\":null,\"custom_image\":null},\"task-add-with-error-a6ed24ad74\":{\"id\":\"task-add-with-error-a6ed24ad74\",\"fn_ref\":{\"function_name\":\"add_with_error\",\"encoded_function\":[\"gASVswQAAAAAAACMF2Nsb3VkcGlja2xlLmNsb3VkcGlja2xllIwOX21ha2VfZnVuY3Rpb26Uk5Qo\\naACMDV9idWlsdGluX3R5cGWUk5SMCENvZGVUeXBllIWUUpQoSwJLAEsASwNLAksDQxaXAGQBZAJ6\\nCwAAfQJ8AHwBegAAAFMAlIw2U2ltdWxhdGVzIGEgdGFzayB3aXRoIGlucHV0cyB0aGF0IHJhaXNl\\ncyBhbiBleGNlcHRpb24ulEsqSwCHlCmMAWGUjAFilIwBX5SHlIx8L1VzZXJzL3NlYmFzdGlhbm1v\\ncmF3aWVjL2dpdC9vcnF1ZXN0cmEtd29ya2Zsb3ctc2RrL3Byb2plY3RzL29ycXVlc3RyYS13b3Jr\\nZmxvdy1ydW50aW1lL3Rlc3RzL2RhdGEvZ2VuZXJhdGVfd29ya2Zsb3dfZGVmcy5weZSMDmFkZF93\\naXRoX2Vycm9ylGgQSzhDFoAA8AgACQuIUYkGgEHgCwyIcYk1gEyUQwCUKSl0lFKUfZQojAtfX3Bh\\nY2thZ2VfX5ROjAhfX25hbWVfX5SMCF9fbWFpbl9flIwIX19maWxlX1+UaA91Tk5OdJRSlIwcY2xv\\ndWRwaWNrbGUuY2xvdWRwaWNrbGVfZmFzdJSMEl9mdW5jdGlvbl9zZXRzdGF0ZZSTlGgbfZQojBdf\\nVGFza0RlZl9fc2RrX3Rhc2tfYm9keZRoG4wHX2ZuX3JlZpSMIG9ycXVlc3RyYS5zZGsuX2NsaWVu\\ndC5fYmFzZS5fZHNslIwRSW5saW5lRnVuY3Rpb25SZWaUk5RoEGgbhpSBlIwIX2ZuX25hbWWUaBCM\\nEF9vdXRwdXRfbWV0YWRhdGGUaCKMElRhc2tPdXRwdXRNZXRhZGF0YZSTlIlLAYaUgZSMC19wYXJh\\nbWV0ZXJzlIwLY29sbGVjdGlvbnOUjAtPcmRlcmVkRGljdJSTlClSlChoC2gijA1UYXNrUGFyYW1l\\ndGVylJOUaAtoIowNUGFyYW1ldGVyS2luZJSTlIwVUE9TSVRJT05BTF9PUl9LRVlXT1JElIWUUpSG\\nlIGUaAxoM2gMaDiGlIGUdYwKX3Jlc291cmNlc5RoIowJUmVzb3VyY2VzlJOUKE5OTk5OdJSBlIwN\\nX2N1c3RvbV9pbWFnZZROjAxfY3VzdG9tX25hbWWUTowTX2RlcGVuZGVuY3lfaW1wb3J0c5ROjB9f\\ndXNlX2RlZmF1bHRfZGVwZW5kZW5jeV9pbXBvcnRzlIiMDl9zb3VyY2VfaW1wb3J0lGgijAxJbmxp\\nbmVJbXBvcnSUk5QpgZSMGl91c2VfZGVmYXVsdF9zb3VyY2VfaW1wb3J0lIiMDF9tYXhfcmV0cmll\\nc5ROjAlfZW52X3ZhcnOUTnV9lChoF2gQjAxfX3F1YWxuYW1lX1+UaBCMD19fYW5ub3RhdGlvbnNf\\nX5R9lIwOX19rd2RlZmF1bHRzX1+UTowMX19kZWZhdWx0c19flE6MCl9fbW9kdWxlX1+UaBiMB19f\\nZG9jX1+UaAmMC19fY2xvc3VyZV9flE6MF19jbG91ZHBpY2tsZV9zdWJtb2R1bGVzlF2UjAtfX2ds\\nb2JhbHNfX5R9lHWGlIZSMC4=\\n\"],\"type\":\"INLINE_FUNCTION_REF\"},\"parameters\":[{\"name\":\"a\",\"kind\":\"POSITIONAL_OR_KEYWORD\"},{\"name\":\"b\",\"kind\":\"POSITIONAL_OR_KEYWORD\"}],\"output_metadata\":{\"is_subscriptable\":false,\"n_outputs\":1},\"source_import_id\":\"inline-import-6\",\"dependency_import_ids\":null,\"resources\":null,\"max_retries\":null,\"custom_image\":null}},\"artifact_nodes\":{\"artifact-0-add\":{\"id\":\"artifact-0-add\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":null},\"artifact-1-add-with-error\":{\"id\":\"artifact-1-add-with-error\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":null},\"artifact-2-add\":{\"id\":\"artifact-2-add\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":null}},\"constant_nodes\":{\"constant-0\":{\"id\":\"constant-0\",\"value\":\"21\",\"serialization_format\":\"JSON\",\"value_preview\":\"21\"},\"constant-1\":{\"id\":\"constant-1\",\"value\":\"37\",\"serialization_format\":\"JSON\",\"value_preview\":\"37\"}},\"secret_nodes\":{},\"task_invocations\":{\"invocation-0-task-add\":{\"id\":\"invocation-0-task-add\",\"task_id\":\"task-add-e573a4e53a\",\"args_ids\":[\"artifact-1-add-with-error\",\"artifact-1-add-with-error\"],\"kwargs_ids\":{},\"output_ids\":[\"artifact-0-add\"],\"resources\":null,\"custom_image\":null,\"env_vars\":null},\"invocation-1-task-add-with-error\":{\"id\":\"invocation-1-task-add-with-error\",\"task_id\":\"task-add-with-error-a6ed24ad74\",\"args_ids\":[\"artifact-2-add\",\"artifact-2-add\"],\"kwargs_ids\":{},\"output_ids\":[\"artifact-1-add-with-error\"],\"resources\":null,\"custom_image\":null,\"env_vars\":null},\"invocation-2-task-add\":{\"id\":\"invocation-2-task-add\",\"task_id\":\"task-add-e573a4e53a\",\"args_ids\":[\"constant-1\",\"constant-0\"],\"kwargs_ids\":{},\"output_ids\":[\"artifact-2-add\"],\"resources\":null,\"custom_image\":null,\"env_vars\":null}},\"output_ids\":[\"artifact-0-add\"],\"data_aggregation\":null,\"metadata\":{\"sdk_version\":{\"original\":\"0.64.0\",\"major\":0,\"minor\":64,\"patch\":0,\"is_prerelease\":false},\"python_version\":{\"original\":\"3.11.6 (main, Nov  6 2023, 13:24:14) [Clang 15.0.0 (clang-1500.0.40.1)]\",\"major\":3,\"minor\":11,\"patch\":6,\"is_prerelease\":false}},\"resources\":null}"
