+++ projects/orquestra-workflow-runtime/tests/data/greet_wf.json
@@ -1,0 +1,1 @@
+"{\"name\":\"greet_wf\",\"fn_ref\":{\"module\":\"__main__\",\"function_name\":\"greet_wf\",\"file_path\":\"generate_workflow_defs.py\",\"line_number\":198,\"type\":\"MODULE_FUNCTION_REF\"},\"imports\":{\"inline-import-8\":{\"id\":\"inline-import-8\",\"type\":\"INLINE_IMPORT\"}},\"tasks\":{\"task-make-greeting-36522f7e75\":{\"id\":\"task-make-greeting-36522f7e75\",\"fn_ref\":{\"function_name\":\"make_greeting\",\"encoded_function\":[\"gASVkAQAAAAAAACMF2Nsb3VkcGlja2xlLmNsb3VkcGlja2xllIwOX21ha2VfZnVuY3Rpb26Uk5Qo\\naACMDV9idWlsdGluX3R5cGWUk5SMCENvZGVUeXBllIWUUpQoSwJLAEsASwJLBEsDQxKXAGQBfACb\\nAGQCfAGbAJ0EUwCUTowHeW9vb29vIJSMBiBmcm9tIJSHlCmMBG5hbWWUjAxjb21wYW55X25hbWWU\\nhpSMfC9Vc2Vycy9zZWJhc3RpYW5tb3Jhd2llYy9naXQvb3JxdWVzdHJhLXdvcmtmbG93LXNkay9w\\ncm9qZWN0cy9vcnF1ZXN0cmEtd29ya2Zsb3ctcnVudGltZS90ZXN0cy9kYXRhL2dlbmVyYXRlX3dv\\ncmtmbG93X2RlZnMucHmUjA1tYWtlX2dyZWV0aW5nlGgQSx9DGIAA4AsvkFTQCy/QCy+gHNALL9AL\\nL9AEL5RDAJQpKXSUUpR9lCiMC19fcGFja2FnZV9flE6MCF9fbmFtZV9flIwIX19tYWluX1+UjAhf\\nX2ZpbGVfX5RoD3VOTk50lFKUjBxjbG91ZHBpY2tsZS5jbG91ZHBpY2tsZV9mYXN0lIwSX2Z1bmN0\\naW9uX3NldHN0YXRllJOUaBt9lCiMF19UYXNrRGVmX19zZGtfdGFza19ib2R5lGgbjAdfZm5fcmVm\\nlIwgb3JxdWVzdHJhLnNkay5fY2xpZW50Ll9iYXNlLl9kc2yUjBFJbmxpbmVGdW5jdGlvblJlZpST\\nlGgQaBuGlIGUjAhfZm5fbmFtZZRoEIwQX291dHB1dF9tZXRhZGF0YZRoIowSVGFza091dHB1dE1l\\ndGFkYXRhlJOUiUsBhpSBlIwLX3BhcmFtZXRlcnOUjAtjb2xsZWN0aW9uc5SMC09yZGVyZWREaWN0\\nlJOUKVKUKGgMaCKMDVRhc2tQYXJhbWV0ZXKUk5RoDGgijA1QYXJhbWV0ZXJLaW5klJOUjBVQT1NJ\\nVElPTkFMX09SX0tFWVdPUkSUhZRSlIaUgZRoDWgzaA1oOIaUgZR1jApfcmVzb3VyY2VzlGgijAlS\\nZXNvdXJjZXOUk5QoTk5OTk50lIGUjA1fY3VzdG9tX2ltYWdllE6MDF9jdXN0b21fbmFtZZROjBNf\\nZGVwZW5kZW5jeV9pbXBvcnRzlE6MH191c2VfZGVmYXVsdF9kZXBlbmRlbmN5X2ltcG9ydHOUiIwO\\nX3NvdXJjZV9pbXBvcnSUaCKMDElubGluZUltcG9ydJSTlCmBlIwaX3VzZV9kZWZhdWx0X3NvdXJj\\nZV9pbXBvcnSUiIwMX21heF9yZXRyaWVzlE6MCV9lbnZfdmFyc5ROdX2UKGgXaBCMDF9fcXVhbG5h\\nbWVfX5RoEIwPX19hbm5vdGF0aW9uc19flH2UjA5fX2t3ZGVmYXVsdHNfX5ROjAxfX2RlZmF1bHRz\\nX1+UTowKX19tb2R1bGVfX5RoGIwHX19kb2NfX5ROjAtfX2Nsb3N1cmVfX5ROjBdfY2xvdWRwaWNr\\nbGVfc3VibW9kdWxlc5RdlIwLX19nbG9iYWxzX1+UfZR1hpSGUjAu\\n\"],\"type\":\"INLINE_FUNCTION_REF\"},\"parameters\":[{\"name\":\"name\",\"kind\":\"POSITIONAL_OR_KEYWORD\"},{\"name\":\"company_name\",\"kind\":\"POSITIONAL_OR_KEYWORD\"}],\"output_metadata\":{\"is_subscriptable\":false,\"n_outputs\":1},\"source_import_id\":\"inline-import-8\",\"dependency_import_ids\":null,\"resources\":null,\"max_retries\":null,\"custom_image\":null}},\"artifact_nodes\":{\"artifact-0-make-greeting\":{\"id\":\"artifact-0-make-greeting\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":null}},\"constant_nodes\":{\"constant-0\":{\"id\":\"constant-0\",\"value\":\"\\\"zapata computing\\\"\",\"serialization_format\":\"JSON\",\"value_preview\":\"'zapata comp\"},\"constant-1\":{\"id\":\"constant-1\",\"value\":\"\\\"emiliano\\\"\",\"serialization_format\":\"JSON\",\"value_preview\":\"'emiliano'\"}},\"secret_nodes\":{},\"task_invocations\":{\"invocation-0-task-make-greeting\":{\"id\":\"invocation-0-task-make-greeting\",\"task_id\":\"task-make-greeting-36522f7e75\",\"args_ids\":[\"constant-1\",\"constant-0\"],\"kwargs_ids\":{},\"output_ids\":[\"artifact-0-make-greeting\"],\"resources\":null,\"custom_image\":null,\"env_vars\":null}},\"output_ids\":[\"artifact-0-make-greeting\"],\"data_aggregation\":null,\"metadata\":{\"sdk_version\":{\"original\":\"0.64.0\",\"major\":0,\"minor\":64,\"patch\":0,\"is_prerelease\":false},\"python_version\":{\"original\":\"3.11.6 (main, Nov  6 2023, 13:24:14) [Clang 15.0.0 (clang-1500.0.40.1)]\",\"major\":3,\"minor\":11,\"patch\":6,\"is_prerelease\":false}},\"resources\":null}"
