+++ projects/orquestra-workflow-runtime/tests/data/infinite_workflow.json
@@ -1,0 +1,1 @@
+"{\"name\":\"infinite_workflow\",\"fn_ref\":{\"module\":\"__main__\",\"function_name\":\"infinite_workflow\",\"file_path\":\"generate_workflow_defs.py\",\"line_number\":272,\"type\":\"MODULE_FUNCTION_REF\"},\"imports\":{\"inline-import-12\":{\"id\":\"inline-import-12\",\"type\":\"INLINE_IMPORT\"}},\"tasks\":{\"task-long-task-1ddd252033\":{\"id\":\"task-long-task-1ddd252033\",\"fn_ref\":{\"function_name\":\"long_task\",\"encoded_function\":[\"gASVwgQAAAAAAACMF2Nsb3VkcGlja2xlLmNsb3VkcGlja2xllIwOX21ha2VfZnVuY3Rpb26Uk5Qo\\naACMDV9idWlsdGluX3R5cGWUk5SMCENvZGVUeXBllIWUUpQoSwBLAEsASwJLA0sXQy6XAGQBZABs\\nAH0BAgB8AWoBAAAAAAAAAABkAqYBAACrAQAAAAAAAAAAAQBkAFMAlE5LAE0QDoeUjAR0aW1llIwF\\nc2xlZXCUhpSMAV+UaAqGlIx8L1VzZXJzL3NlYmFzdGlhbm1vcmF3aWVjL2dpdC9vcnF1ZXN0cmEt\\nd29ya2Zsb3ctc2RrL3Byb2plY3RzL29ycXVlc3RyYS13b3JrZmxvdy1ydW50aW1lL3Rlc3RzL2Rh\\ndGEvZ2VuZXJhdGVfd29ya2Zsb3dfZGVmcy5weZSMCWxvbmdfdGFza5SMLWdlbmVyYXRlX2luZmlu\\naXRlX3dvcmtmbG93Ljxsb2NhbHM+LmxvbmdfdGFza5RNCQFDJYAA4AgTiAuIC4gL8AYACROIBIwK\\nkDfRCBvUCBvQCBvQCBvQCBuUQwCUKSl0lFKUfZQojAtfX3BhY2thZ2VfX5ROjAhfX25hbWVfX5SM\\nCF9fbWFpbl9flIwIX19maWxlX1+UaA91Tk5OdJRSlIwcY2xvdWRwaWNrbGUuY2xvdWRwaWNrbGVf\\nZmFzdJSMEl9mdW5jdGlvbl9zZXRzdGF0ZZSTlGgcfZQojBdfVGFza0RlZl9fc2RrX3Rhc2tfYm9k\\neZRoHIwHX2ZuX3JlZpSMIG9ycXVlc3RyYS5zZGsuX2NsaWVudC5fYmFzZS5fZHNslIwRSW5saW5l\\nRnVuY3Rpb25SZWaUk5RoEGgchpSBlIwIX2ZuX25hbWWUaBCMEF9vdXRwdXRfbWV0YWRhdGGUaCOM\\nElRhc2tPdXRwdXRNZXRhZGF0YZSTlIlLAYaUgZSMC19wYXJhbWV0ZXJzlIwLY29sbGVjdGlvbnOU\\njAtPcmRlcmVkRGljdJSTlClSlGgNaCOMDVRhc2tQYXJhbWV0ZXKUk5RoDWgjjA1QYXJhbWV0ZXJL\\naW5klJOUjA5WQVJfUE9TSVRJT05BTJSFlFKUhpSBlHOMCl9yZXNvdXJjZXOUaCOMCVJlc291cmNl\\nc5STlChOTk5OTnSUgZSMDV9jdXN0b21faW1hZ2WUTowMX2N1c3RvbV9uYW1llE6ME19kZXBlbmRl\\nbmN5X2ltcG9ydHOUTowfX3VzZV9kZWZhdWx0X2RlcGVuZGVuY3lfaW1wb3J0c5SIjA5fc291cmNl\\nX2ltcG9ydJRoI4wMSW5saW5lSW1wb3J0lJOUKYGUjBpfdXNlX2RlZmF1bHRfc291cmNlX2ltcG9y\\ndJSIjAxfbWF4X3JldHJpZXOUTowJX2Vudl92YXJzlE51fZQoaBhoEIwMX19xdWFsbmFtZV9flGgR\\njA9fX2Fubm90YXRpb25zX1+UfZSMDl9fa3dkZWZhdWx0c19flE6MDF9fZGVmYXVsdHNfX5ROjApf\\nX21vZHVsZV9flGgZjAdfX2RvY19flE6MC19fY2xvc3VyZV9flE6MF19jbG91ZHBpY2tsZV9zdWJt\\nb2R1bGVzlF2UjAtfX2dsb2JhbHNfX5R9lHWGlIZSMC4=\\n\"],\"type\":\"INLINE_FUNCTION_REF\"},\"parameters\":[{\"name\":\"_\",\"kind\":\"VAR_POSITIONAL\"}],\"output_metadata\":{\"is_subscriptable\":false,\"n_outputs\":1},\"source_import_id\":\"inline-import-12\",\"dependency_import_ids\":null,\"resources\":null,\"max_retries\":null,\"custom_image\":null}},\"artifact_nodes\":{\"artifact-0-long-task\":{\"id\":\"artifact-0-long-task\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":null}},\"constant_nodes\":{},\"secret_nodes\":{},\"task_invocations\":{\"invocation-0-task-long-task\":{\"id\":\"invocation-0-task-long-task\",\"task_id\":\"task-long-task-1ddd252033\",\"args_ids\":[],\"kwargs_ids\":{},\"output_ids\":[\"artifact-0-long-task\"],\"resources\":null,\"custom_image\":null,\"env_vars\":null}},\"output_ids\":[\"artifact-0-long-task\"],\"data_aggregation\":null,\"metadata\":{\"sdk_version\":{\"original\":\"0.64.0\",\"major\":0,\"minor\":64,\"patch\":0,\"is_prerelease\":false},\"python_version\":{\"original\":\"3.11.6 (main, Nov  6 2023, 13:24:14) [Clang 15.0.0 (clang-1500.0.40.1)]\",\"major\":3,\"minor\":11,\"patch\":6,\"is_prerelease\":false}},\"resources\":null}"
