+++ projects/orquestra-workflow-runtime/tests/data/multioutput_task_wf.json
@@ -1,0 +1,1 @@
+"{\"name\":\"multioutput_task_wf\",\"fn_ref\":{\"module\":\"__main__\",\"function_name\":\"multioutput_task_wf\",\"file_path\":\"generate_workflow_defs.py\",\"line_number\":184,\"type\":\"MODULE_FUNCTION_REF\"},\"imports\":{\"inline-import-7\":{\"id\":\"inline-import-7\",\"type\":\"INLINE_IMPORT\"}},\"tasks\":{\"task-multioutput-task-81d106a29f\":{\"id\":\"task-multioutput-task-81d106a29f\",\"fn_ref\":{\"function_name\":\"multioutput_task\",\"encoded_function\":[\"gASVCgQAAAAAAACMF2Nsb3VkcGlja2xlLmNsb3VkcGlja2xllIwOX21ha2VfZnVuY3Rpb26Uk5Qo\\naACMDV9idWlsdGluX3R5cGWUk5SMCENvZGVUeXBllIWUUpQoSwBLAEsASwBLAUsDQwaXAGQBUwCU\\nTowGWmFwYXRhlIwJQ29tcHV0aW5nlIaUhpQpKYx8L1VzZXJzL3NlYmFzdGlhbm1vcmF3aWVjL2dp\\ndC9vcnF1ZXN0cmEtd29ya2Zsb3ctc2RrL3Byb2plY3RzL29ycXVlc3RyYS13b3JrZmxvdy1ydW50\\naW1lL3Rlc3RzL2RhdGEvZ2VuZXJhdGVfd29ya2Zsb3dfZGVmcy5weZSMEG11bHRpb3V0cHV0X3Rh\\nc2uUaA5LJEMIgADgCyDQCyCUQwCUKSl0lFKUfZQojAtfX3BhY2thZ2VfX5ROjAhfX25hbWVfX5SM\\nCF9fbWFpbl9flIwIX19maWxlX1+UaA11Tk5OdJRSlIwcY2xvdWRwaWNrbGUuY2xvdWRwaWNrbGVf\\nZmFzdJSMEl9mdW5jdGlvbl9zZXRzdGF0ZZSTlGgZfZQojBdfVGFza0RlZl9fc2RrX3Rhc2tfYm9k\\neZRoGYwHX2ZuX3JlZpSMIG9ycXVlc3RyYS5zZGsuX2NsaWVudC5fYmFzZS5fZHNslIwRSW5saW5l\\nRnVuY3Rpb25SZWaUk5RoDmgZhpSBlIwIX2ZuX25hbWWUaA6MEF9vdXRwdXRfbWV0YWRhdGGUaCCM\\nElRhc2tPdXRwdXRNZXRhZGF0YZSTlIhLAoaUgZSMC19wYXJhbWV0ZXJzlIwLY29sbGVjdGlvbnOU\\njAtPcmRlcmVkRGljdJSTlClSlIwKX3Jlc291cmNlc5RoIIwJUmVzb3VyY2VzlJOUKE5OTk5OdJSB\\nlIwNX2N1c3RvbV9pbWFnZZROjAxfY3VzdG9tX25hbWWUTowTX2RlcGVuZGVuY3lfaW1wb3J0c5RO\\njB9fdXNlX2RlZmF1bHRfZGVwZW5kZW5jeV9pbXBvcnRzlIiMDl9zb3VyY2VfaW1wb3J0lGggjAxJ\\nbmxpbmVJbXBvcnSUk5QpgZSMGl91c2VfZGVmYXVsdF9zb3VyY2VfaW1wb3J0lIiMDF9tYXhfcmV0\\ncmllc5ROjAlfZW52X3ZhcnOUTnV9lChoFWgOjAxfX3F1YWxuYW1lX1+UaA6MD19fYW5ub3RhdGlv\\nbnNfX5R9lIwOX19rd2RlZmF1bHRzX1+UTowMX19kZWZhdWx0c19flE6MCl9fbW9kdWxlX1+UaBaM\\nB19fZG9jX1+UTowLX19jbG9zdXJlX1+UTowXX2Nsb3VkcGlja2xlX3N1Ym1vZHVsZXOUXZSMC19f\\nZ2xvYmFsc19flH2UdYaUhlIwLg==\\n\"],\"type\":\"INLINE_FUNCTION_REF\"},\"parameters\":[],\"output_metadata\":{\"is_subscriptable\":true,\"n_outputs\":2},\"source_import_id\":\"inline-import-7\",\"dependency_import_ids\":null,\"resources\":null,\"max_retries\":null,\"custom_image\":null}},\"artifact_nodes\":{\"artifact-0-multioutput-task\":{\"id\":\"artifact-0-multioutput-task\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":0},\"artifact-1-multioutput-task\":{\"id\":\"artifact-1-multioutput-task\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":1},\"artifact-3-multioutput-task\":{\"id\":\"artifact-3-multioutput-task\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":0},\"artifact-4-multioutput-task\":{\"id\":\"artifact-4-multioutput-task\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":1},\"artifact-6-multioutput-task\":{\"id\":\"artifact-6-multioutput-task\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":0},\"artifact-7-multioutput-task\":{\"id\":\"artifact-7-multioutput-task\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":1},\"artifact-9-multioutput-task\":{\"id\":\"artifact-9-multioutput-task\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":0},\"artifact-10-multioutput-task\":{\"id\":\"artifact-10-multioutput-task\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":1},\"artifact-2-multioutput-task\":{\"id\":\"artifact-2-multioutput-task\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":null},\"artifact-5-multioutput-task\":{\"id\":\"artifact-5-multioutput-task\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":null},\"artifact-8-multioutput-task\":{\"id\":\"artifact-8-multioutput-task\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":null},\"artifact-11-multioutput-task\":{\"id\":\"artifact-11-multioutput-task\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":null}},\"constant_nodes\":{},\"secret_nodes\":{},\"task_invocations\":{\"invocation-0-task-multioutput-task\":{\"id\":\"invocation-0-task-multioutput-task\",\"task_id\":\"task-multioutput-task-81d106a29f\",\"args_ids\":[],\"kwargs_ids\":{},\"output_ids\":[\"artifact-0-multioutput-task\",\"artifact-1-multioutput-task\",\"artifact-2-multioutput-task\"],\"resources\":null,\"custom_image\":null,\"env_vars\":null},\"invocation-1-task-multioutput-task\":{\"id\":\"invocation-1-task-multioutput-task\",\"task_id\":\"task-multioutput-task-81d106a29f\",\"args_ids\":[],\"kwargs_ids\":{},\"output_ids\":[\"artifact-3-multioutput-task\",\"artifact-4-multioutput-task\",\"artifact-5-multioutput-task\"],\"resources\":null,\"custom_image\":null,\"env_vars\":null},\"invocation-2-task-multioutput-task\":{\"id\":\"invocation-2-task-multioutput-task\",\"task_id\":\"task-multioutput-task-81d106a29f\",\"args_ids\":[],\"kwargs_ids\":{},\"output_ids\":[\"artifact-6-multioutput-task\",\"artifact-7-multioutput-task\",\"artifact-8-multioutput-task\"],\"resources\":null,\"custom_image\":null,\"env_vars\":null},\"invocation-3-task-multioutput-task\":{\"id\":\"invocation-3-task-multioutput-task\",\"task_id\":\"task-multioutput-task-81d106a29f\",\"args_ids\":[],\"kwargs_ids\":{},\"output_ids\":[\"artifact-9-multioutput-task\",\"artifact-10-multioutput-task\",\"artifact-11-multioutput-task\"],\"resources\":null,\"custom_image\":null,\"env_vars\":null}},\"output_ids\":[\"artifact-9-multioutput-task\",\"artifact-10-multioutput-task\",\"artifact-7-multioutput-task\",\"artifact-3-multioutput-task\",\"artifact-2-multioutput-task\",\"artifact-0-multioutput-task\",\"artifact-1-multioutput-task\"],\"data_aggregation\":null,\"metadata\":{\"sdk_version\":{\"original\":\"0.64.0\",\"major\":0,\"minor\":64,\"patch\":0,\"is_prerelease\":false},\"python_version\":{\"original\":\"3.11.6 (main, Nov  6 2023, 13:24:14) [Clang 15.0.0 (clang-1500.0.40.1)]\",\"major\":3,\"minor\":11,\"patch\":6,\"is_prerelease\":false}},\"resources\":null}"
