+++ projects/orquestra-workflow-runtime/tests/data/serial_wf_with_slow_middle_task.json
@@ -1,0 +1,1 @@
+"{\"name\":\"serial_wf_with_slow_middle_task\",\"fn_ref\":{\"module\":\"__main__\",\"function_name\":\"serial_wf_with_slow_middle_task\",\"file_path\":\"generate_workflow_defs.py\",\"line_number\":224,\"type\":\"MODULE_FUNCTION_REF\"},\"imports\":{\"inline-import-10\":{\"id\":\"inline-import-10\",\"type\":\"INLINE_IMPORT\"}},\"tasks\":{\"task-add-56b7cc98a0\":{\"id\":\"task-add-56b7cc98a0\",\"fn_ref\":{\"function_name\":\"add\",\"encoded_function\":[\"gASVUgQAAAAAAACMF2Nsb3VkcGlja2xlLmNsb3VkcGlja2xllIwOX21ha2VfZnVuY3Rpb26Uk5Qo\\naACMDV9idWlsdGluX3R5cGWUk5SMCENvZGVUeXBllIWUUpQoSwJLAEsASwJLAksDQwyXAHwAfAF6\\nAAAAUwCUToWUKYwBYZSMAWKUhpSMfC9Vc2Vycy9zZWJhc3RpYW5tb3Jhd2llYy9naXQvb3JxdWVz\\ndHJhLXdvcmtmbG93LXNkay9wcm9qZWN0cy9vcnF1ZXN0cmEtd29ya2Zsb3ctcnVudGltZS90ZXN0\\ncy9kYXRhL2dlbmVyYXRlX3dvcmtmbG93X2RlZnMucHmUjANhZGSUaA5LQUMLgADgCwyIcYk1gEyU\\nQwCUKSl0lFKUfZQojAtfX3BhY2thZ2VfX5ROjAhfX25hbWVfX5SMCF9fbWFpbl9flIwIX19maWxl\\nX1+UaA11Tk5OdJRSlIwcY2xvdWRwaWNrbGUuY2xvdWRwaWNrbGVfZmFzdJSMEl9mdW5jdGlvbl9z\\nZXRzdGF0ZZSTlGgZfZQojBdfVGFza0RlZl9fc2RrX3Rhc2tfYm9keZRoGYwHX2ZuX3JlZpSMIG9y\\ncXVlc3RyYS5zZGsuX2NsaWVudC5fYmFzZS5fZHNslIwRSW5saW5lRnVuY3Rpb25SZWaUk5RoDmgZ\\nhpSBlIwIX2ZuX25hbWWUaA6MEF9vdXRwdXRfbWV0YWRhdGGUaCCMElRhc2tPdXRwdXRNZXRhZGF0\\nYZSTlIlLAYaUgZSMC19wYXJhbWV0ZXJzlIwLY29sbGVjdGlvbnOUjAtPcmRlcmVkRGljdJSTlClS\\nlChoCmggjA1UYXNrUGFyYW1ldGVylJOUaApoIIwNUGFyYW1ldGVyS2luZJSTlIwVUE9TSVRJT05B\\nTF9PUl9LRVlXT1JElIWUUpSGlIGUaAtoMWgLaDaGlIGUdYwKX3Jlc291cmNlc5RoIIwJUmVzb3Vy\\nY2VzlJOUKE5OTk5OdJSBlIwNX2N1c3RvbV9pbWFnZZROjAxfY3VzdG9tX25hbWWUTowTX2RlcGVu\\nZGVuY3lfaW1wb3J0c5ROjB9fdXNlX2RlZmF1bHRfZGVwZW5kZW5jeV9pbXBvcnRzlIiMDl9zb3Vy\\nY2VfaW1wb3J0lGggjAxJbmxpbmVJbXBvcnSUk5QpgZSMGl91c2VfZGVmYXVsdF9zb3VyY2VfaW1w\\nb3J0lIiMDF9tYXhfcmV0cmllc5ROjAlfZW52X3ZhcnOUTnV9lChoFWgOjAxfX3F1YWxuYW1lX1+U\\naA6MD19fYW5ub3RhdGlvbnNfX5R9lIwOX19rd2RlZmF1bHRzX1+UTowMX19kZWZhdWx0c19flE6M\\nCl9fbW9kdWxlX1+UaBaMB19fZG9jX1+UTowLX19jbG9zdXJlX1+UTowXX2Nsb3VkcGlja2xlX3N1\\nYm1vZHVsZXOUXZSMC19fZ2xvYmFsc19flH2UdYaUhlIwLg==\\n\"],\"type\":\"INLINE_FUNCTION_REF\"},\"parameters\":[{\"name\":\"a\",\"kind\":\"POSITIONAL_OR_KEYWORD\"},{\"name\":\"b\",\"kind\":\"POSITIONAL_OR_KEYWORD\"}],\"output_metadata\":{\"is_subscriptable\":false,\"n_outputs\":1},\"source_import_id\":\"inline-import-10\",\"dependency_import_ids\":null,\"resources\":null,\"max_retries\":null,\"custom_image\":null},\"task-add-slow-022e2f34ba\":{\"id\":\"task-add-slow-022e2f34ba\",\"fn_ref\":{\"function_name\":\"add_slow\",\"encoded_function\":[\"gASVtwQAAAAAAACMF2Nsb3VkcGlja2xlLmNsb3VkcGlja2xllIwOX21ha2VfZnVuY3Rpb26Uk5Qo\\naACMDV9idWlsdGluX3R5cGWUk5SMCENvZGVUeXBllIWUUpQoSwJLAEsASwJLA0sDQzSXAHQBAAAA\\nAAAAAAAAAGoBAAAAAAAAAABkAaYBAACrAQAAAAAAAAAAAQB8AHwBegAAAFMAlE5LCoaUjAR0aW1l\\nlIwFc2xlZXCUhpSMAWGUjAFilIaUjHwvVXNlcnMvc2ViYXN0aWFubW9yYXdpZWMvZ2l0L29ycXVl\\nc3RyYS13b3JrZmxvdy1zZGsvcHJvamVjdHMvb3JxdWVzdHJhLXdvcmtmbG93LXJ1bnRpbWUvdGVz\\ndHMvZGF0YS9nZW5lcmF0ZV93b3JrZmxvd19kZWZzLnB5lIwIYWRkX3Nsb3eUaBFLDEMYgADlBAiE\\nSohygU6EToBO4AsMiHGJNYBMlEMAlCkpdJRSlH2UKIwLX19wYWNrYWdlX1+UTowIX19uYW1lX1+U\\njAhfX21haW5fX5SMCF9fZmlsZV9flGgQdU5OTnSUUpSMHGNsb3VkcGlja2xlLmNsb3VkcGlja2xl\\nX2Zhc3SUjBJfZnVuY3Rpb25fc2V0c3RhdGWUk5RoHH2UKIwXX1Rhc2tEZWZfX3Nka190YXNrX2Jv\\nZHmUaByMB19mbl9yZWaUjCBvcnF1ZXN0cmEuc2RrLl9jbGllbnQuX2Jhc2UuX2RzbJSMEUlubGlu\\nZUZ1bmN0aW9uUmVmlJOUaBFoHIaUgZSMCF9mbl9uYW1llGgRjBBfb3V0cHV0X21ldGFkYXRhlGgj\\njBJUYXNrT3V0cHV0TWV0YWRhdGGUk5SJSwGGlIGUjAtfcGFyYW1ldGVyc5SMC2NvbGxlY3Rpb25z\\nlIwLT3JkZXJlZERpY3SUk5QpUpQoaA1oI4wNVGFza1BhcmFtZXRlcpSTlGgNaCOMDVBhcmFtZXRl\\ncktpbmSUk5SMFVBPU0lUSU9OQUxfT1JfS0VZV09SRJSFlFKUhpSBlGgOaDRoDmg5hpSBlHWMCl9y\\nZXNvdXJjZXOUaCOMCVJlc291cmNlc5STlChOTk5OTnSUgZSMDV9jdXN0b21faW1hZ2WUTowMX2N1\\nc3RvbV9uYW1llE6ME19kZXBlbmRlbmN5X2ltcG9ydHOUTowfX3VzZV9kZWZhdWx0X2RlcGVuZGVu\\nY3lfaW1wb3J0c5SIjA5fc291cmNlX2ltcG9ydJRoI4wMSW5saW5lSW1wb3J0lJOUKYGUjBpfdXNl\\nX2RlZmF1bHRfc291cmNlX2ltcG9ydJSIjAxfbWF4X3JldHJpZXOUTowJX2Vudl92YXJzlE51fZQo\\naBhoEYwMX19xdWFsbmFtZV9flGgRjA9fX2Fubm90YXRpb25zX1+UfZSMDl9fa3dkZWZhdWx0c19f\\nlE6MDF9fZGVmYXVsdHNfX5ROjApfX21vZHVsZV9flGgZjAdfX2RvY19flE6MC19fY2xvc3VyZV9f\\nlE6MF19jbG91ZHBpY2tsZV9zdWJtb2R1bGVzlF2UjAtfX2dsb2JhbHNfX5R9lGgKaACMCXN1Ymlt\\ncG9ydJSTlGgKhZRSlHN1hpSGUjAu\\n\"],\"type\":\"INLINE_FUNCTION_REF\"},\"parameters\":[{\"name\":\"a\",\"kind\":\"POSITIONAL_OR_KEYWORD\"},{\"name\":\"b\",\"kind\":\"POSITIONAL_OR_KEYWORD\"}],\"output_metadata\":{\"is_subscriptable\":false,\"n_outputs\":1},\"source_import_id\":\"inline-import-10\",\"dependency_import_ids\":null,\"resources\":null,\"max_retries\":null,\"custom_image\":null}},\"artifact_nodes\":{\"artifact-0-add\":{\"id\":\"artifact-0-add\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":null},\"artifact-1-add-slow\":{\"id\":\"artifact-1-add-slow\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":null},\"artifact-2-add\":{\"id\":\"artifact-2-add\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":null}},\"constant_nodes\":{\"constant-0\":{\"id\":\"constant-0\",\"value\":\"37\",\"serialization_format\":\"JSON\",\"value_preview\":\"37\"},\"constant-1\":{\"id\":\"constant-1\",\"value\":\"21\",\"serialization_format\":\"JSON\",\"value_preview\":\"21\"}},\"secret_nodes\":{},\"task_invocations\":{\"invocation-0-task-add\":{\"id\":\"invocation-0-task-add\",\"task_id\":\"task-add-56b7cc98a0\",\"args_ids\":[\"artifact-1-add-slow\",\"artifact-1-add-slow\"],\"kwargs_ids\":{},\"output_ids\":[\"artifact-0-add\"],\"resources\":null,\"custom_image\":null,\"env_vars\":null},\"invocation-1-task-add-slow\":{\"id\":\"invocation-1-task-add-slow\",\"task_id\":\"task-add-slow-022e2f34ba\",\"args_ids\":[\"artifact-2-add\",\"artifact-2-add\"],\"kwargs_ids\":{},\"output_ids\":[\"artifact-1-add-slow\"],\"resources\":null,\"custom_image\":null,\"env_vars\":null},\"invocation-2-task-add\":{\"id\":\"invocation-2-task-add\",\"task_id\":\"task-add-56b7cc98a0\",\"args_ids\":[\"constant-1\",\"constant-0\"],\"kwargs_ids\":{},\"output_ids\":[\"artifact-2-add\"],\"resources\":null,\"custom_image\":null,\"env_vars\":null}},\"output_ids\":[\"artifact-0-add\"],\"data_aggregation\":null,\"metadata\":{\"sdk_version\":{\"original\":\"0.64.0\",\"major\":0,\"minor\":64,\"patch\":0,\"is_prerelease\":false},\"python_version\":{\"original\":\"3.11.6 (main, Nov  6 2023, 13:24:14) [Clang 15.0.0 (clang-1500.0.40.1)]\",\"major\":3,\"minor\":11,\"patch\":6,\"is_prerelease\":false}},\"resources\":null}"
