+++ projects/orquestra-workflow-runtime/tests/data/simple_wf.json
@@ -1,0 +1,1 @@
+"{\"name\":\"hello_orquestra_wf\",\"fn_ref\":{\"module\":\"__main__\",\"function_name\":\"hello_orquestra_wf\",\"file_path\":\"generate_workflow_defs.py\",\"line_number\":77,\"type\":\"MODULE_FUNCTION_REF\"},\"imports\":{\"inline-import-2\":{\"id\":\"inline-import-2\",\"type\":\"INLINE_IMPORT\"}},\"tasks\":{\"task-hello-orquestra-3df0bcd3dd\":{\"id\":\"task-hello-orquestra-3df0bcd3dd\",\"fn_ref\":{\"function_name\":\"hello_orquestra\",\"encoded_function\":[\"gASVWwQAAAAAAACMF2Nsb3VkcGlja2xlLmNsb3VkcGlja2xllIwOX21ha2VfZnVuY3Rpb26Uk5Qo\\naACMDV9idWlsdGluX3R5cGWUk5SMCENvZGVUeXBllIWUUpQoSwBLAEsASwBLAUsTQwaXAGQBUwCU\\nTowQSGVsbG8gT3JxdWVzdHJhIZSGlCkpjHwvVXNlcnMvc2ViYXN0aWFubW9yYXdpZWMvZ2l0L29y\\ncXVlc3RyYS13b3JrZmxvdy1zZGsvcHJvamVjdHMvb3JxdWVzdHJhLXdvcmtmbG93LXJ1bnRpbWUv\\ndGVzdHMvZGF0YS9nZW5lcmF0ZV93b3JrZmxvd19kZWZzLnB5lIwPaGVsbG9fb3JxdWVzdHJhlIw4\\nbWFrZV93b3JrZmxvd193aXRoX2RlcGVuZGVuY2llcy48bG9jYWxzPi5oZWxsb19vcnF1ZXN0cmGU\\nS0lDCIAA4A8h0A8hlEMAlCkpdJRSlH2UKIwLX19wYWNrYWdlX1+UTowIX19uYW1lX1+UjAhfX21h\\naW5fX5SMCF9fZmlsZV9flGgLdU5OTnSUUpSMHGNsb3VkcGlja2xlLmNsb3VkcGlja2xlX2Zhc3SU\\njBJfZnVuY3Rpb25fc2V0c3RhdGWUk5RoGH2UKIwXX1Rhc2tEZWZfX3Nka190YXNrX2JvZHmUaBiM\\nB19mbl9yZWaUjCBvcnF1ZXN0cmEuc2RrLl9jbGllbnQuX2Jhc2UuX2RzbJSMEUlubGluZUZ1bmN0\\naW9uUmVmlJOUaAxoGIaUgZSMCF9mbl9uYW1llGgMjBBfb3V0cHV0X21ldGFkYXRhlGgfjBJUYXNr\\nT3V0cHV0TWV0YWRhdGGUk5SJSwGGlIGUjAtfcGFyYW1ldGVyc5SMC2NvbGxlY3Rpb25zlIwLT3Jk\\nZXJlZERpY3SUk5QpUpSMCl9yZXNvdXJjZXOUaB+MCVJlc291cmNlc5STlChOTk5OTnSUgZSMDV9j\\ndXN0b21faW1hZ2WUTowMX2N1c3RvbV9uYW1llE6ME19kZXBlbmRlbmN5X2ltcG9ydHOUKYwfX3Vz\\nZV9kZWZhdWx0X2RlcGVuZGVuY3lfaW1wb3J0c5SJjA5fc291cmNlX2ltcG9ydJRoH4wMSW5saW5l\\nSW1wb3J0lJOUKYGUjBpfdXNlX2RlZmF1bHRfc291cmNlX2ltcG9ydJSIjAxfbWF4X3JldHJpZXOU\\nTowJX2Vudl92YXJzlE51fZQoaBRoDIwMX19xdWFsbmFtZV9flGgNjA9fX2Fubm90YXRpb25zX1+U\\nfZSMBnJldHVybpSMCGJ1aWx0aW5zlIwDc3RylJOUc4wOX19rd2RlZmF1bHRzX1+UTowMX19kZWZh\\ndWx0c19flE6MCl9fbW9kdWxlX1+UaBWMB19fZG9jX1+UTowLX19jbG9zdXJlX1+UTowXX2Nsb3Vk\\ncGlja2xlX3N1Ym1vZHVsZXOUXZSMC19fZ2xvYmFsc19flH2UdYaUhlIwLg==\\n\"],\"type\":\"INLINE_FUNCTION_REF\"},\"parameters\":[],\"output_metadata\":{\"is_subscriptable\":false,\"n_outputs\":1},\"source_import_id\":\"inline-import-2\",\"dependency_import_ids\":[],\"resources\":null,\"max_retries\":null,\"custom_image\":null}},\"artifact_nodes\":{\"artifact-0-hello-orquestra\":{\"id\":\"artifact-0-hello-orquestra\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":null}},\"constant_nodes\":{},\"secret_nodes\":{},\"task_invocations\":{\"invocation-0-task-hello-orquestra\":{\"id\":\"invocation-0-task-hello-orquestra\",\"task_id\":\"task-hello-orquestra-3df0bcd3dd\",\"args_ids\":[],\"kwargs_ids\":{},\"output_ids\":[\"artifact-0-hello-orquestra\"],\"resources\":null,\"custom_image\":null,\"env_vars\":null}},\"output_ids\":[\"artifact-0-hello-orquestra\"],\"data_aggregation\":null,\"metadata\":{\"sdk_version\":{\"original\":\"0.64.0\",\"major\":0,\"minor\":64,\"patch\":0,\"is_prerelease\":false},\"python_version\":{\"original\":\"3.11.6 (main, Nov  6 2023, 13:24:14) [Clang 15.0.0 (clang-1500.0.40.1)]\",\"major\":3,\"minor\":11,\"patch\":6,\"is_prerelease\":false}},\"resources\":null}"
