+++ projects/orquestra-workflow-runtime/tests/data/wf_with_exec_ctx.json
@@ -1,0 +1,1 @@
+"{\"name\":\"wf_with_exec_ctx\",\"fn_ref\":{\"module\":\"__main__\",\"function_name\":\"wf_with_exec_ctx\",\"file_path\":\"generate_workflow_defs.py\",\"line_number\":175,\"type\":\"MODULE_FUNCTION_REF\"},\"imports\":{\"inline-import-5\":{\"id\":\"inline-import-5\",\"type\":\"INLINE_IMPORT\"}},\"tasks\":{\"task-get-exec-ctx-75f54a13fa\":{\"id\":\"task-get-exec-ctx-75f54a13fa\",\"fn_ref\":{\"function_name\":\"get_exec_ctx\",\"encoded_function\":[\"gASVKQUAAAAAAACMF2Nsb3VkcGlja2xlLmNsb3VkcGlja2xllIwOX21ha2VfZnVuY3Rpb26Uk5Qo\\naACMDV9idWlsdGluX3R5cGWUk5SMCENvZGVUeXBllIWUUpQoSwBLAEsASwJLAksTQ1SXAGQBZABs\\nAH0AfABqAQAAAAAAAAAAagIAAAAAAAAAAKADAAAAAAAAAAAAAAAAAAAAAAAAAACmAAAAqwAAAAAA\\nAAAAAH0BfAFqBAAAAAAAAAAAUwCUTksAhpQojCJvcnF1ZXN0cmEud29ya2Zsb3dfc2hhcmVkLmV4\\nZWNfY3R4lIwPd29ya2Zsb3dfc2hhcmVklIwIZXhlY19jdHiUjBhnZXRfY3VycmVudF9leGVjX2Nv\\nbnRleHSUjARuYW1llHSUjAlvcnF1ZXN0cmGUjANjdHiUhpSMfC9Vc2Vycy9zZWJhc3RpYW5tb3Jh\\nd2llYy9naXQvb3JxdWVzdHJhLXdvcmtmbG93LXNkay9wcm9qZWN0cy9vcnF1ZXN0cmEtd29ya2Zs\\nb3ctcnVudGltZS90ZXN0cy9kYXRhL2dlbmVyYXRlX3dvcmtmbG93X2RlZnMucHmUjAxnZXRfZXhl\\nY19jdHiUjC9nZW5lcmF0ZV93Zl93aXRoX2V4ZWNfY3R4Ljxsb2NhbHM+LmdldF9leGVjX2N0eJRL\\nqEMsgADgCDHQCDHQCDHQCDHgDhfUDifUDjDXDknSDknRDkvUDkuIA9gPEox4iA+UQwCUKSl0lFKU\\nfZQojAtfX3BhY2thZ2VfX5ROjAhfX25hbWVfX5SMCF9fbWFpbl9flIwIX19maWxlX1+UaBN1Tk5O\\ndJRSlIwcY2xvdWRwaWNrbGUuY2xvdWRwaWNrbGVfZmFzdJSMEl9mdW5jdGlvbl9zZXRzdGF0ZZST\\nlGggfZQojBdfVGFza0RlZl9fc2RrX3Rhc2tfYm9keZRoIIwHX2ZuX3JlZpSMIG9ycXVlc3RyYS5z\\nZGsuX2NsaWVudC5fYmFzZS5fZHNslIwRSW5saW5lRnVuY3Rpb25SZWaUk5RoFGgghpSBlIwIX2Zu\\nX25hbWWUaBSMEF9vdXRwdXRfbWV0YWRhdGGUaCeMElRhc2tPdXRwdXRNZXRhZGF0YZSTlIlLAYaU\\ngZSMC19wYXJhbWV0ZXJzlIwLY29sbGVjdGlvbnOUjAtPcmRlcmVkRGljdJSTlClSlIwKX3Jlc291\\ncmNlc5RoJ4wJUmVzb3VyY2VzlJOUKE5OTk5OdJSBlIwNX2N1c3RvbV9pbWFnZZROjAxfY3VzdG9t\\nX25hbWWUTowTX2RlcGVuZGVuY3lfaW1wb3J0c5ROjB9fdXNlX2RlZmF1bHRfZGVwZW5kZW5jeV9p\\nbXBvcnRzlIiMDl9zb3VyY2VfaW1wb3J0lGgnjAxJbmxpbmVJbXBvcnSUk5QpgZSMGl91c2VfZGVm\\nYXVsdF9zb3VyY2VfaW1wb3J0lIiMDF9tYXhfcmV0cmllc5ROjAlfZW52X3ZhcnOUTnV9lChoHGgU\\njAxfX3F1YWxuYW1lX1+UaBWMD19fYW5ub3RhdGlvbnNfX5R9lIwGcmV0dXJulIwIYnVpbHRpbnOU\\njANzdHKUk5RzjA5fX2t3ZGVmYXVsdHNfX5ROjAxfX2RlZmF1bHRzX1+UTowKX19tb2R1bGVfX5Ro\\nHYwHX19kb2NfX5ROjAtfX2Nsb3N1cmVfX5ROjBdfY2xvdWRwaWNrbGVfc3VibW9kdWxlc5RdlIwL\\nX19nbG9iYWxzX1+UfZR1hpSGUjAu\\n\"],\"type\":\"INLINE_FUNCTION_REF\"},\"parameters\":[],\"output_metadata\":{\"is_subscriptable\":false,\"n_outputs\":1},\"source_import_id\":\"inline-import-5\",\"dependency_import_ids\":null,\"resources\":null,\"max_retries\":null,\"custom_image\":null}},\"artifact_nodes\":{\"artifact-0-get-exec-ctx\":{\"id\":\"artifact-0-get-exec-ctx\",\"custom_name\":null,\"serialization_format\":\"AUTO\",\"artifact_index\":null}},\"constant_nodes\":{},\"secret_nodes\":{},\"task_invocations\":{\"invocation-0-task-get-exec-ctx\":{\"id\":\"invocation-0-task-get-exec-ctx\",\"task_id\":\"task-get-exec-ctx-75f54a13fa\",\"args_ids\":[],\"kwargs_ids\":{},\"output_ids\":[\"artifact-0-get-exec-ctx\"],\"resources\":null,\"custom_image\":null,\"env_vars\":null}},\"output_ids\":[\"artifact-0-get-exec-ctx\"],\"data_aggregation\":null,\"metadata\":{\"sdk_version\":{\"original\":\"0.64.0\",\"major\":0,\"minor\":64,\"patch\":0,\"is_prerelease\":false},\"python_version\":{\"original\":\"3.11.6 (main, Nov  6 2023, 13:24:14) [Clang 15.0.0 (clang-1500.0.40.1)]\",\"major\":3,\"minor\":11,\"patch\":6,\"is_prerelease\":false}},\"resources\":null}"
