+++ projects/orquestra-workflow-runtime/tests/test_build_workflow.py
@@ -1,13 +1,17 @@
+import json
+from pathlib import Path
 from unittest.mock import Mock, call, create_autospec
 
 import orquestra.workflow_shared.secrets
+import pydantic
 import pytest
 from orquestra.workflow_shared import parse_git_url, serde
+from orquestra.workflow_shared._graphs import iter_invocations_topologically
 from orquestra.workflow_shared.schema import ir
-from orquestra.workflow_shared.schema.ir import GitURL, SecretNode
+from orquestra.workflow_shared.schema.ir import GitURL, SecretNode, WorkflowDef
 from orquestra.workflow_shared.schema.responses import WorkflowResult
 
-from orquestra.workflow_runtime._ray import _build_workflow
+from orquestra.workflow_runtime._ray import _build_workflow, _client
 
 
 @pytest.fixture
@@ -432,3 +436,208 @@ class TestArgumentUnwrapper:
 
             calls = [call(packed), call(unpacked[0]), call(unpacked[1])]
             mock_deserialize.assert_has_calls(calls)
+
+
+class TestMakeDag:
+    @pytest.fixture
+    def client(self):
+        return create_autospec(_client.RayClient)
+
+    @pytest.fixture
+    def wf_run_id(self):
+        return "mocked_wf_run_id"
+
+    def test_import_pip_string_resolved_once_per_import(
+        self, monkeypatch: pytest.MonkeyPatch, client: Mock, wf_run_id: str
+    ):
+        import orquestra.workflow_shared.packaging._versions
+
+        monkeypatch.setattr(
+            orquestra.workflow_shared.packaging._versions,
+            "get_installed_version",
+            lambda _: "0.64.0",
+        )
+        monkeypatch.setattr(
+            orquestra.workflow_runtime._ray._build_workflow,
+            "get_installed_version",
+            lambda _: "0.64.0",
+        )
+        pip_string = create_autospec(_build_workflow._pip_string)
+        pip_string.return_value = ["mocked"]
+        monkeypatch.setattr(_build_workflow, "_pip_string", pip_string)
+        path_to_json = Path(__file__).parent.joinpath("data/100_tasks_with_imps.json")
+        with open(path_to_json, "r", encoding="utf-8") as f:
+            wf_def = pydantic.TypeAdapter(WorkflowDef).validate_json(json.load(f))
+        _ = _build_workflow.make_ray_dag(client, wf_def, wf_run_id, False)
+        assert pip_string.mock_calls == [call(imp) for imp in wf_def.imports.values()]
+
+
+@pytest.mark.parametrize(
+    "installed_sdk_version, expected_sdk_dependency",
+    [
+        ("0.57.1.dev3+g3ef9f57.d20231003", None),
+        ("1.2.3", "1.2.3"),
+        ("0.1.dev1+g25df81e", None),
+    ],
+)
+class TestHandlingSDKVersions:
+    """``_import_pip_env`` handles adding the current SDK version as a dependency.
+
+    Note that these tests don't mock serde - we're interested in whether the correct
+    package list gets constructed so we can't just return 'mocked' for imports.
+    """
+
+    @staticmethod
+    def test_with_no_dependencies(
+        installed_sdk_version: str,
+        expected_sdk_dependency: str,
+        monkeypatch: pytest.MonkeyPatch,
+    ):
+        import orquestra.workflow_shared.packaging._versions
+
+        # Given
+        monkeypatch.setattr(
+            _build_workflow,
+            "get_installed_version",
+            Mock(return_value=installed_sdk_version),
+        )
+        monkeypatch.setattr(
+            orquestra.workflow_shared.packaging._versions,
+            "get_installed_version",
+            Mock(return_value=installed_sdk_version),
+        )
+        path_to_json = Path(__file__).parent.joinpath("data/simple_wf.json")
+
+        with open(path_to_json, "r", encoding="utf-8") as f:
+            wf = pydantic.TypeAdapter(WorkflowDef).validate_json(json.load(f))
+        task_inv = [inv for inv in iter_invocations_topologically(wf)][0]
+        imports = {
+            id: _build_workflow._pip_string(imp) for id, imp in wf.imports.items()
+        }
+
+        # When
+        pip = _build_workflow._import_pip_env(task_inv, wf, imports)
+
+        # Then
+        if expected_sdk_dependency:
+            assert pip == [f"orquestra-sdk=={expected_sdk_dependency}"]
+        else:
+            assert pip == []
+
+    @staticmethod
+    def test_with_multiple_dependencies(
+        installed_sdk_version: str,
+        expected_sdk_dependency: str,
+        monkeypatch: pytest.MonkeyPatch,
+    ):
+        # Given
+        python_import = "MarkupSafe"
+        import_version = "==1.0.0"
+        import orquestra.workflow_shared.packaging._versions
+
+        monkeypatch.setattr(
+            _build_workflow,
+            "get_installed_version",
+            Mock(return_value=installed_sdk_version),
+        )
+        monkeypatch.setattr(
+            orquestra.workflow_shared.packaging._versions,
+            "get_installed_version",
+            Mock(return_value=installed_sdk_version),
+        )
+        path_to_json = Path(__file__).parent.joinpath("data/wildcard.json")
+
+        with open(path_to_json, "r", encoding="utf-8") as f:
+            wildcard_workflow_def = json.load(f)
+            wildcard_workflow_def = wildcard_workflow_def.replace(
+                "<WILDCARD_NAME>", python_import
+            )
+            wildcard_workflow_def = wildcard_workflow_def.replace(
+                "<WILDCARD_VERSION>", import_version
+            )
+            wf = pydantic.TypeAdapter(WorkflowDef).validate_json(wildcard_workflow_def)
+
+        task_inv = [inv for inv in iter_invocations_topologically(wf)][0]
+        imports = {
+            id: _build_workflow._pip_string(imp) for id, imp in wf.imports.items()
+        }
+
+        # When
+        pip = _build_workflow._import_pip_env(task_inv, wf, imports)
+
+        # Then
+        if expected_sdk_dependency:
+            assert sorted(pip) == sorted(
+                [f"{python_import}{import_version}"]
+                + [f"orquestra-sdk=={expected_sdk_dependency}"]
+            )
+        else:
+            assert sorted(pip) == sorted([f"{python_import}{import_version}"])
+
+    @pytest.mark.parametrize(
+        "sdk_import",
+        [
+            "orquestra-sdk",
+            "orquestra-sdk>=1.2.3",
+            "orquestra-sdk<=1.2.3",
+            "orquestra-sdk~=1.2.3",
+            "orquestra-sdk==1.2.3",
+            "orquestra-sdk!=1.2.3",
+            "orquestra-sdk===1.2.3",
+            "orquestra-sdk >= 1.2.3",
+            "orquestra-sdk <= 1.2.3",
+            "orquestra-sdk ~= 1.2.3",
+            "orquestra-sdk == 1.2.3",
+            "orquestra-sdk != 1.2.3",
+            "orquestra-sdk === 1.2.3",
+        ],
+    )
+    class TestHandlesSDKDependency:
+        @staticmethod
+        @pytest.mark.filterwarnings("ignore:The definition for task ")
+        def test_replaces_declared_sdk_dependency(
+            sdk_import,
+            installed_sdk_version: str,
+            expected_sdk_dependency: str,
+            monkeypatch: pytest.MonkeyPatch,
+        ):
+            # Given
+            sdk_version = sdk_import.replace("orquestra-sdk", "").strip()
+            import orquestra.workflow_shared.packaging._versions
+
+            monkeypatch.setattr(
+                _build_workflow,
+                "get_installed_version",
+                Mock(return_value=installed_sdk_version),
+            )
+            monkeypatch.setattr(
+                orquestra.workflow_shared.packaging._versions,
+                "get_installed_version",
+                Mock(return_value=installed_sdk_version),
+            )
+            path_to_json = Path(__file__).parent.joinpath("data/wildcard.json")
+            with open(path_to_json, "r", encoding="utf-8") as f:
+                wildcard_workflow_def = json.load(f)
+                wildcard_workflow_def = wildcard_workflow_def.replace(
+                    "<WILDCARD_NAME>", "orquestra-sdk"
+                )
+                wildcard_workflow_def = wildcard_workflow_def.replace(
+                    "<WILDCARD_VERSION>", sdk_version
+                )
+                wf = pydantic.TypeAdapter(WorkflowDef).validate_json(
+                    wildcard_workflow_def
+                )
+            task_inv = [inv for inv in iter_invocations_topologically(wf)][0]
+            imports = {
+                id: _build_workflow._pip_string(imp) for id, imp in wf.imports.items()
+            }
+
+            # When
+            pip = _build_workflow._import_pip_env(task_inv, wf, imports)
+
+            # Then
+            if expected_sdk_dependency:
+                assert pip == [f"orquestra-sdk=={expected_sdk_dependency}"]
+
+            else:
+                assert pip == []
